% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ctd.R
\name{add_ctd}
\alias{add_ctd}
\title{Add CellTypeDataset}
\usage{
add_ctd(
  results = load_example_results(),
  ctd = load_example_ctd(),
  annotLevel = length(ctd),
  keep_specificity_quantiles = NULL,
  keep_mean_exp_quantiles = NULL,
  keep_celltypes = NULL,
  rep_dt = NULL,
  all.x = FALSE,
  allow.cartesian = TRUE,
  by = c("gene_symbol", "CellType"),
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{ctd}{Cell Type Data List generated using
\link[EWCE]{generate_celltype_data}.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{keep_specificity_quantiles}{Which cell type
specificity quantiles to keep (max quantile is 40).}

\item{keep_mean_exp_quantiles}{Which cell type
mean expression quantiles to keep (max quantile is 40).}

\item{keep_celltypes}{A character vector of cell types to keep.}

\item{rep_dt}{Report table.}

\item{all.x}{logical; if \code{TRUE}, rows from \code{x} which have no matching row
in \code{y} are included. These rows will have 'NA's in the columns that are usually
filled with values from \code{y}. The default is \code{FALSE} so that only rows with
data from both \code{x} and \code{y} are included in the output.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{verbose}{Print messages.}
}
\description{
Annotate genes in \code{results} with data from a CellTypeDataset (CTD).
The following columns will be added:
\itemize{
\item{"specificity"} Cell-type specificity score.
\item{"specificity_quantiles"} Cell-type specificity quantile.
\item{"mean_exp"} Mean expression per cell-type.
\item{"mean_exp_quantiles"} Mean expression quantile per cell-type.
}
}
\examples{
results <- load_example_results()[seq(100),]
results2 <- add_ctd(results=results)
}
