% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_driver_genes.R
\name{add_driver_genes}
\alias{add_driver_genes}
\title{Add driver genes}
\usage{
add_driver_genes(
  results = load_example_results(),
  ctd_list = load_example_ctd(file = paste0("ctd_", unique(results$ctd), ".rds"),
    multi_dataset = TRUE),
  annotLevels = map_ctd_levels(results),
  keep_quantiles = seq(30, 40),
  min_value = NULL,
  metric = "specificity_quantiles",
  top_n = NULL,
  group_var = "hpo_id",
  celltype_var = "CellType",
  ...
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{ctd_list}{A named list of CellTypeDataset objects each
created with \link[EWCE]{generate_celltype_data}.}

\item{annotLevels}{The annotation level to use within each CTD in
\code{ctd_list}.}

\item{keep_quantiles}{Quantiles to keep in each CellTypeDataset of the
\code{ctd_list}.}

\item{min_value}{Miniumum specificity quantile to keep.}

\item{metric}{Which metric to use in the CellTypeDatasets.}

\item{top_n}{Top N genes to keep when grouping by \code{group_vars}.}

\item{group_var}{Grouping variable to use when selecting top N genes.
Only used when \code{top_n!=NULL}.}

\item{celltype_var}{The name of the cell type column to merge on.}

\item{...}{
  Arguments passed on to \code{\link[HPOExplorer:add_]{HPOExplorer::add_genes}}
  \describe{
    \item{\code{phenos}}{A data.table containing HPO IDs and other metadata.}
    \item{\code{hpo}}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}
    \item{\code{all.x}}{logical; if \code{TRUE}, rows from \code{x} which have no matching row
in \code{y} are included. These rows will have 'NA's in the columns that are usually
filled with values from \code{y}. The default is \code{FALSE} so that only rows with
data from both \code{x} and \code{y} are included in the output.}
    \item{\code{allow.cartesian}}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}
    \item{\code{phenotype_to_genes}}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}
    \item{\code{by}}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}
    \item{\code{gene_col}}{Name of the gene column.}
  }}
}
\description{
This function identifies "driver genes" from phenotype-cell type association analyses.
For a given phenotype-cell type pair, driver genes are defined as the intersect of genes that had a
phenotype evidence score >0 and were within the top 75th expression specificity percentile
(quantiles 30-40 out of 40) for the associated cell type.
}
\examples{
res <- load_example_results()[seq(100)]
res <- add_driver_genes(results=res)
}
