% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fold_change.R
\name{add_logfc}
\alias{add_logfc}
\title{Add log fold-change}
\usage{
add_logfc(results, effect_var = "estimate", force_new = FALSE)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{effect_var}{Name of the effect size column in the \code{results}.}

\item{force_new}{Force the creation of a new \code{logFC} column
even when one already exists.}
}
\description{
Compute \code{log2(fold-change)} based on the effect size.
}
\examples{
results = load_example_results()[seq(5000)]
add_logfc(results)
}
