% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symptom_results.R
\name{add_symptom_results}
\alias{add_symptom_results}
\title{Add symptom results}
\usage{
add_symptom_results(
  results = load_example_results(multi_dataset = TRUE),
  q_threshold = 0.05,
  fold_threshold = 2,
  celltype_col = "CellType",
  ctd_list = load_example_ctd(file = paste0("ctd_", unique(results$ctd), ".rds"),
    multi_dataset = TRUE),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  annotLevels = list(DescartesHuman = 2, HumanCellLandscape = 3),
  keep_quantiles = seq(30, 40),
  top_n = NULL,
  proportion_driver_genes_symptom_threshold = 0.75
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{top_n}{Top N genes to keep when grouping by \code{group_vars}.}
}
\description{
Add symptom results to the results data.table.
}
\examples{
results <- add_symptom_results()
}
