% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anatomy_maps}
\alias{anatomy_maps}
\title{Anatomy maps}
\format{
data.table
}
\source{
\code{
## Download standardised Seurat objects from CellXGene.
obj_list <- list()
obj_list[["DescartesHuman"]] <- readRDS("~/Downloads/db4d63ab-0ca3-4fa5-b0cf-34955227912d.rds")
obj_list[["HumanCellLandscape"]] <- readRDS("~/Downloads/1b7484e3-83a0-47fe-847e-54d811a2adae.rds")
cols <- c("tissue_ontology_term_id",
          "tissue",
          # "tissue_original",
          "cell_type_ontology_term_id",
          "cell_type")
anatomy_maps <- lapply(obj_list, function(obj){
  select_cols <- cols[cols %in% colnames(obj@meta.data)]
  unique(data.table::as.data.table(obj@meta.data[,select_cols]))
}) |> data.table::rbindlist(idcol = "ctd", fill=TRUE)
usethis::use_data(anatomy_maps, overwrite = TRUE)
}
}
\usage{
data("anatomy_maps")
}
\description{
Mappings from cell-type names to anatomy ontology terms in the
HumanCellLandscape and DescartesHuman scRNA-seq datasets.
}
\keyword{datasets}
