% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_heatmap.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{Plot heatmap}
\usage{
correlation_heatmap(
  top_targets,
  row_side_vars = c("ancestor_name", "ontLvl", "Onset_top"),
  col_side_vars = c("n_celltypes", "n_genes"),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  col = pals::gnuplot(),
  show_plot = TRUE,
  save_plot = tempfile(fileext = "correlation_heatmap.pdf"),
  height = 10,
  width = height * 1.3,
  fontsize = 7,
  row_km = 3,
  column_km = row_km,
  row_km_repeats = 1000,
  column_km_repeats = row_km_repeats,
  seed = 2023,
  interact = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{top_targets}{\link[data.table]{data.table} of prioritised targets
generated by \link[MultiEWCE]{prioritise_targets}.}

\item{row_side_vars}{Variables to include in row-side
metadata annotations.}

\item{col_side_vars}{Variables to include in column-side
metadata annotations.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{show_plot}{Print the plot after generating it.}

\item{save_plot}{Path to save plot to.}

\item{height}{Height of the heatmap body.}

\item{width}{Width of the heatmap body.}

\item{fontsize}{Axis labels font size.}

\item{row_km}{Same as \code{km}.}

\item{column_km}{K-means clustering on columns.}

\item{row_km_repeats}{Number of k-means runs to get a consensus k-means clustering. Note if \code{row_km_repeats} is set to more than one, the final number of groups might be smaller than \code{row_km}, but this might means the original \code{row_km} is not a good choice.}

\item{column_km_repeats}{Number of k-means runs to get a consensus k-means clustering. Similar as \code{row_km_repeats}.}

\item{seed}{Set the seed for reproducible clustering.}

\item{interact}{Make the plot interactive via \link[heatmaply]{heatmaply}.}

\item{verbose}{Print messages.}
}
\value{
Plot
}
\description{
Plot a phenotype x phenotype correlation matrix based on genetic overlap.
}
\examples{
top_targets <- example_targets$top_targets
hm <- correlation_heatmap(top_targets = top_targets)
}
