% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_results.R
\name{count_results}
\alias{count_results}
\title{Count results}
\usage{
count_results(
  results,
  q_threshold = 5e-04,
  fold_threshold = 1,
  sd_threshold = NULL,
  p_threshold = NULL,
  group_var = "CellType",
  order = -1
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{sd_threshold}{The standard deviations from the mean
threshold to subset the results by.}

\item{p_threshold}{The p-value
threshold to subset the results by.}

\item{group_var}{Variable to group results by.}

\item{order}{ An integer vector with only possible values of \code{1} and
\code{-1}, corresponding to ascending and descending order. The length of
\code{order} must be either \code{1} or equal to that of \code{cols}. If
\code{length(order) == 1}, it is recycled to \code{length(cols)}. }
}
\value{
data.table with counts.
}
\description{
Count the number of test that remain per celltype (or phenotype)
after filtering by one or more criterion.
}
\examples{
results <- load_example_results()
celltype_counts <- count_results(results=results, group_var="CellType")
phenotype_counts <- count_results(results=results, group_var="Phenotype")
}
