% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dt.R
\name{create_dt}
\alias{create_dt}
\title{Interactive DT}
\usage{
create_dt(
  dat,
  caption = "",
  buttons = c("copy", "csv", "excel", "pdf", "print"),
  scrollY = 400
)
}
\arguments{
\item{dat}{Data to show.}

\item{caption}{Table caption.}

\item{buttons}{Download button types to include in table.}

\item{scrollY}{Maximum height (in pixels) before defaulting to scrolling.}
}
\description{
Generate an interactive data table with download buttons.
}
\examples{
create_dt(dat = mtcars)
}
\concept{general}
