% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_targets}
\alias{example_targets}
\title{Example prioritised targets}
\format{
ontology_index
}
\source{
\code{
example_targets <- prioritise_targets()
usethis::use_data(example_targets, overwrite = TRUE)
}
}
\usage{
data("example_targets")
}
\description{
Example output from the function \link[MultiEWCE]{prioritise_targets}.
Used in examples to reduce run time.
}
\keyword{datasets}
