% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_help.R
\name{extract_help}
\alias{extract_help}
\title{Extract help}
\source{
\href{https://stackoverflow.com/a/9195691}{Original source.}
}
\usage{
extract_help(pkg, fn = NULL, to = c("txt", "html", "latex", "ex"))
}
\arguments{
\item{pkg}{Package name.}

\item{fn}{Function name,}
}
\value{
text. html, latex formats
}
\description{
Extract help documentation from a function.
}
\keyword{internal}
