% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ggstatsplot_subtitle.R
\name{filter_ggstatsplot_subtitle}
\alias{filter_ggstatsplot_subtitle}
\title{Filter ggstatsplot subtitle to show selected statistics only}
\usage{
filter_ggstatsplot_subtitle(p, stats_idx = c(1, 3, 4, 7))
}
\arguments{
\item{p}{ggstatsplot object}

\item{stats_idx}{Indices of statistics to keep in subtitle}
}
\value{
ggstatsplot object with filtered subtitle
}
\description{
Filter ggstatsplot subtitle to show selected statistics only
}
\examples{
  p <- ggstatsplot::ggscatterstats(
    data = mtcars,
    x = disp,
    y = mpg
  )
  p_filtered <- filter_ggstatsplot_subtitle(p, stats_idx = c(1,3))
  print(p_filtered)
}
