% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_histogram.R
\name{frequency_histogram}
\alias{frequency_histogram}
\title{Gene and phenotype frequencies histogram}
\usage{
frequency_histogram(
  results = load_example_results(),
  phenotype_to_genes = load_phenotype_to_genes(),
  show_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{show_plot}{Print the plot after it has been created.}

\item{verbose}{Print messages.}
}
\value{
ggplot object
}
\description{
Plot the frequency of gene-phenotype and phenotype-disease associations.
}
\examples{
fp_res <- frequency_histogram()
}
