% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_results.R
\name{gen_results}
\alias{gen_results}
\title{Generate results}
\usage{
gen_results(
  ctd,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  list_names = unique(gene_data[[list_name_column]]),
  bg = unique(gene_data[[gene_column]]),
  overwrite_past_analysis = FALSE,
  reps = 100,
  annotLevel = 1,
  genelistSpecies = "human",
  sctSpecies = "human",
  cores = 1,
  save_dir_tmp = NULL,
  save_dir = tempdir(),
  verbose = 1
)
}
\arguments{
\item{ctd}{Cell Type Data List generated using
\link[EWCE]{generate_celltype_data}.}

\item{gene_data}{data frame of gene list names and genes
(see \link[HPOExplorer]{get_gene_lists}).}

\item{list_name_column}{The name of the gene_data column
that has the gene list names.}

\item{gene_column}{The name of the gene_data column that contains the genes.}

\item{list_names}{character vector of gene list names.}

\item{bg}{List of gene symbols containing the background gene list
(including hit genes). If \code{bg=NULL},
 an appropriate gene background will be created automatically.}

\item{overwrite_past_analysis}{overwrite previous results
in the results dir \code{bool}.}

\item{reps}{Number of random gene lists to generate (\emph{Default: 100},
but should be >=10,000 for publication-quality results).}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{genelistSpecies}{Species that \code{hits} genes came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{sctSpecies}{Species that \code{sct_data} came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{cores}{The number of cores to run in parallel (e.g. 8) \code{int}.}

\item{save_dir_tmp}{Folder to save intermediate results files to
(one file per gene list). Set to \code{NULL} to skip saving temporary files.}

\item{save_dir}{Folder to save merged results in.}

\item{verbose}{Print messages.}
}
\value{
All results as a dataframe.
}
\description{
Generates EWCE results on multiple gene lists in parallel by calling
\code{ewce_para}. It allows you to stop the analysis and then continue later
from where you left off as it checks the results output directory for
finished gene lists and removes them from the input.
It also excludes gene lists with
less than 4 unique genes (which cause errors in EWCE analysis).
}
\details{
The gene_data should be a data frame that contains a column of gene
list names (e.g. the column may be called "Phenotype"), and a column of
genes (e.g. "Gene"). For example:\tabular{ll}{
   Phenotype \tab Gene \cr
   "Abnormal heart" \tab gene X \cr
   "Abnormal heart" \tab gene Y \cr
   "Poor vision" \tab gene Z \cr
   "Poor vision" \tab gene Y \cr
   etc... \tab  \cr
}


For more information on this see docs for get_gene_list
(\link[HPOExplorer]{get_gene_lists}).
}
\examples{
gene_data <- HPOExplorer::load_phenotype_to_genes()
list_names <- unique(gene_data$Phenotype)[seq_len(5)]
ctd <- load_example_ctd()
all_results <- gen_results(ctd = ctd,
                           gene_data = gene_data,
                           list_names = list_names,
                           reps = 10)
}
