% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_ontology.R
\name{get_cell_ontology}
\alias{get_cell_ontology}
\title{Subset EWCE results}
\usage{
get_cell_ontology(
  cell_type,
  results = load_example_results(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes()
)
}
\arguments{
\item{cell_type}{The cell type of interest to be plotted.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}
}
\value{
A data frame of the selected subset of RD EWCE results
with HPO ID column added.
}
\description{
Subset RD EWCE results data by cell type, fold change and q value.
}
\examples{
signif_cell_data <- get_cell_ontology(cell_type="Amacrine cells")
}
