% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cl.R
\name{get_cl}
\alias{get_cl}
\title{Get Cell Ontology}
\usage{
get_cl(name = "cl", tag = "v2023-09-21", lvl = 1, remove_rings = TRUE, ...)
}
\arguments{
\item{name}{\describe{
\item{<...>}{Any ontology name from \link[KGExplorer]{get_ols_options}}
\item{"mondo"}{
Import the \href{https://mondo.monarchinitiative.org/}{Mondo} ontology.
\href{https://github.com/monarch-initiative/mondo/release}{
GitHub.}}.
\item{"hpo"}{
Import the \href{https://hpo.jax.org/app/}{Human Phenotype Ontology}.
\href{https://github.com/obophenotype/human-phenotype-ontology/release}{
GitHub.}}
}}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{lvl}{Depth of the ancestor terms to add. 
Will get the closest ancestor to this level if none have this exact level.}

\item{remove_rings}{Parameter passed to \code{simona::import_ontology}.}

\item{...}{
  Arguments passed on to \code{\link[KGExplorer:get_]{KGExplorer::get_ontology}}
  \describe{
    \item{\code{save_dir}}{Directory to save a file to.}
    \item{\code{force_new}}{If TRUE, force a new download.}
    \item{\code{method}}{Compute ontology levels using:
\describe{
 \item{"height" (default)}{\link[simona]{dag_height}.}
 \item{"depth"}{\link[simona]{dag_depth}.}
}}
    \item{\code{filetype}}{File type to search for.}
    \item{\code{import_func}}{Function to import the ontology with. 
If \code{NULL}, automatically tries to choose the correct function.}
    \item{\code{terms}}{A vector of ontology term IDs.}
    \item{\code{add_metadata}}{Add metadata to the resulting ontology object.}
    \item{\code{add_n_edges}}{Add the number of edges (connections) for each term.}
    \item{\code{add_ontology_levels}}{Add the ontology level for each term.}
  }}
}
\description{
A thin wrapper around \link[KGExplorer]{get_ontology} to get the
specific version of the Cell Ontology used in the original analyses that
produced the results stored in \link{load_example_results}.
}
\examples{
cl <- get_cl()
}
