% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get remote data}
\usage{
get_data(
  fname,
  repo = "neurogenomics/MSTExplorer",
  save_dir = KGExplorer::cache_dir(package = "MSTExplorer"),
  overwrite = TRUE,
  tag = "latest",
  check = FALSE
)
}
\arguments{
\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{save_dir}{Directory to save data to.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}

\item{tag}{tag for the GitHub release to which this data should be attached.}
}
\description{
Download remotely stored data via \link[piggyback]{pb_download}.
}
\keyword{internal}
