% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_list.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Get gene list}
\usage{
get_gene_list(
  list_name,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene"
)
}
\arguments{
\item{list_name}{Names of one or more gene lists.}

\item{gene_data}{data frame of gene list names and genes
(see \link[MultiEWCE]{get_gene_list}).}

\item{list_name_column}{The name of the gene_data column
that has the gene list names.}

\item{gene_column}{The name of the gene_data column that contains the genes.}
}
\value{
A character vector of genes associated with the selected list_name
}
\description{
This returns a character vector of genes (a gene list). The gene_data should
be a data frame that contains a column of gene list names (e.g. the column
may be called "Phenotype"), and a column of genes (e.g. "Gene"). For example,
lets call the following data frame \code{phenotype_to_genes}:
}
\details{
\tabular{ll}{
   Phenotype \tab Gene \cr
   "Abnormal heart" \tab gene X \cr
   "Abnormal heart" \tab gene Y \cr
   "Poor vision" \tab gene Z \cr
   "Poor vision" \tab gene Y \cr
}
}
\examples{
gene_data <- HPOExplorer::load_phenotype_to_genes()
gene_list <- get_gene_list(list_name="Focal non-motor seizure",
                           gene_data = gene_data)
}
