% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_replication_rate.R
\name{get_replication_rate}
\alias{get_replication_rate}
\title{Calculate replication rate between two studies}
\usage{
get_replication_rate(dat, x, y, alpha = 0.05)
}
\arguments{
\item{dat}{A data.table containing p-values from two studies.}

\item{x}{The column name of the first study's p-values.}

\item{y}{The column name of the second study's p-values.}

\item{alpha}{Significance threshold for determining replication.}
}
\value{
A list with replication rates from study A to B, B to A, and symmetric replication rate.
}
\description{
Calculate replication rate between two studies
}
\examples{
dat <- data.table::data.table(
  q_DescartesHuman = c(0.01, 0.2, 0.03, 0.5, 0.04),
  q_HumanCellLandscape = c(0.02, 0.03, 0.2, 0.6, 0.01)
)
replication_rates <- get_replication_rate(dat,
x="q_DescartesHuman", y="q_HumanCellLandscape",
)
print(replication_rates)
}
