% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_target_branches.R
\name{get_target_branches}
\alias{get_target_branches}
\title{Get target branches}
\usage{
get_target_branches()
}
\description{
A named list of mappings between HPO ancestral terms (list names),
and Cell Ontology ancestral terms (list values).
}
\keyword{internal}
