% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unfinished_list_names.R
\name{get_unfinished_list_names}
\alias{get_unfinished_list_names}
\title{Get unfinished list names}
\usage{
get_unfinished_list_names(list_names, results_dir)
}
\arguments{
\item{list_names}{A char vector of gene list names}

\item{results_dir}{The directory containing analysed results.}
}
\value{
A character vector of list_names that still need to be analysed.
}
\description{
Gets the unfinished gene list names. It reads the file names in the results
directory and uses this to deduce which gene lists have already been analysed.
This means you can pause the analysis of multiple gene lists and it will not
re-analyse the already completed ones when you start again.
}
\examples{
\dontrun{
list_names <- unique(data$phenotypes)
results_dir <- "output/"
get_unfinished_list_names(list_names,results_dir)
}
}
