% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_gene_lists.R
\name{get_valid_gene_lists}
\alias{get_valid_gene_lists}
\title{get valid gene lists}
\usage{
get_valid_gene_lists(
  ctd,
  list_names,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene"
)
}
\arguments{
\item{ctd}{CTD (cell type data file) see EWCE docs}

\item{list_names}{Vector of gene list names \<character vector\>}

\item{gene_data}{data frame containing a gene column and a column of list_names}

\item{list_name_column}{The name of the column that contains
gene list_names \<string\>}

\item{gene_column}{The name of the gene column (e.g. "Gene") \<string\>}
}
\value{
A character vector of list_names that are associated with a valid
number of genes
}
\description{
Gene lists for EWCE must have 4 or more genes. These genes must also be present
in the CTD file (see EWCE docs), and must be unique (occasionally a gene list
may contain the same gene repeated. This would cause an error).
}
\details{
This function gets gene lists that have atleast four unique genes that are also
present in the CTD file.
}
\examples{
\dontrun{
remove_invalid_gene_lists(ctd,
                          list_names,
                          gene_data,
                          list_names_column = "Phenotype",
                          gene_column = "Gene")
}
}
