% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork_plot_full.R
\name{ggnetwork_plot_full}
\alias{ggnetwork_plot_full}
\title{Create interactive network plot start to finish}
\usage{
ggnetwork_plot_full(
  cell_type,
  results = load_example_results(),
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  columns = list(Phenotype = "Phenotype", ID = "HPO_ID", ontLvl_genes =
    "ontLvl_geneCount_ratio", Description = "description", CellType = "CellType", p =
    "p", q = "q", fold_change = "fold_change"),
  colour_column = "fold_change",
  interactive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cell_type}{The cell type of interest to be plotted.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[HPOExplorer]{make_hoverboxes}.}

\item{colour_column}{column name of the variable to be mapped
to colour \\<string\\>}

\item{interactive}{Make the plot interactive with \link[plotly]{ggplotly}.}

\item{verbose}{Print messages.}
}
\value{
A named list of outputs,
including a interactive network plot of the selected subset
of results from RD EWCE analysis.
}
\description{
This puts all the functions together from getting the subset of results
to creating the final interactive plot.
}
\examples{
res_list <- ggnetwork_plot_full(cell_type = "Microglia")
}
