% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_to_matrix.R
\name{hpo_to_matrix}
\alias{hpo_to_matrix}
\title{HPO to matrix}
\usage{
hpo_to_matrix(
  terms = NULL,
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  formula = "Gene ~ Phenotype",
  fun.aggregate = mean,
  fill = 0,
  run_cor = FALSE,
  as_matrix = TRUE,
  sparse = FALSE,
  method = "pearson",
  verbose = TRUE
)
}
\arguments{
\item{terms}{A subset of HPO IDs to include.
Set to \code{NULL} (default) to include all terms.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{formula}{A formula of the form LHS ~ RHS to cast, see Details.}

\item{fun.aggregate}{Should the data be aggregated before casting? If the formula doesn't identify a single observation for each cell, then aggregation defaults to \code{length} with a message.

  To use multiple aggregation functions, pass a \code{list}; see Examples. }

\item{fill}{Value with which to fill missing cells. If \code{fun.aggregate} is present, takes the value by applying the function on a 0-length vector.}

\item{run_cor}{Return a matrix of pairwise correlations.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{verbose}{Print messages.}
}
\value{
A gene x phenotype matrix,
or a phenotype x phenotype matrix if \code{run_cor=TRUE}.
}
\description{
Convert the HPO to a gene x phenotype matrix.
}
\keyword{internal}
