% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_not_analysed.R
\name{is_not_analysed}
\alias{is_not_analysed}
\title{Is not analysed?}
\usage{
is_not_analysed(list_name, results_dir)
}
\arguments{
\item{list_name}{The name of a gene list e.g. "Phenotypic abnormality"}

\item{results_dir}{The path to results directory}
}
\value{
True or false \code{bool}
}
\description{
Takes a gene list name and checks the output results directory to see if that
gene list has been analysed yet. The ewce_para function outputs the results
for each gene list in to the results directory with name format list_name.rds
}
