% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_ctd.R
\name{load_example_ctd}
\alias{load_example_ctd}
\title{Load example CTD (Descartes)}
\usage{
load_example_ctd(
  file = "CTD_Descartes_example.rds",
  tag = "latest",
  save_dir = tools::R_user_dir(package = "MultiEWCE")
)
}
\arguments{
\item{file}{Name of a remotely stored file.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{save_dir}{Where to store the file locally.}
}
\description{
This loads a example CTD for vignettes and testing purposes.
It is \href{https://descartes.brotmanbaty.org/}{Descartes} Gene Expression
During Development single-cell RNA-seq atlas at annotation level 1.
}
\examples{
CTD <- load_example_ctd()
}
