% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_results.R
\name{load_example_results}
\alias{load_example_results}
\title{Load example results}
\source{
\code{
d <-  "~/Desktop/ewce/rare_disease_celltyping_apps/cell_select_interactive"
piggyback::pb_upload(file = file.path(d,"data/Descartes_All_Results.rds"),
                     tag = "v0.0.1", repo = "neurogenomics/MultiEWCE")
piggyback::pb_upload(file = file.path(d,"data/tabulamuris_merged.rds"),
                     tag = "v0.0.1", repo = "neurogenomics/MultiEWCE")
}
}
\usage{
load_example_results(
  file = "Descartes_All_Results.rds",
  tag = "v0.0.1",
  save_dir = tools::R_user_dir(package = "MultiEWCE")
)
}
\arguments{
\item{file}{File to load:
\itemize{
\item{"Descartes_All_Results.rds":}{
Used the \href{https://descartes.brotmanbaty.org/}{Descartes}
CTD ( annotation level 1).
}
\item{"tabulamuris_merged.rds":}{
Used the \href{https://tabula-muris.ds.czbiohub.org/}{Tabula Muris}
CTD.
}
}}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{save_dir}{Where to store the file locally.}
}
\value{
dataframe of enrichment results.
}
\description{
This loads a example of enrichment results from \link[MultiEWCE]{gen_results}
using the Human Phenotype Ontology and a given CellTypeDataset (CTD).
Results were then merged together with  \link[MultiEWCE]{merge_results}.
}
\examples{
res <- load_example_results()
}
