% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_hpo_graph.R
\name{load_hpo_graph}
\alias{load_hpo_graph}
\title{Load HPO graph}
\source{
\code{
phenos <- HPOExplorer::make_phenos_dataframe(ancestor="All")
g <- HPOExplorer::make_igraph_object(phenos = phenos)
f <- file.path(tempdir(),"hpo_igraph.rds")
saveRDS(g,file = f)
piggyback::pb_upload(file = f,
                     tag = "v0.0.1", repo = "neurogenomics/MultiEWCE")
}
}
\usage{
load_hpo_graph(
  file = "hpo_igraph.rds",
  tag = "latest",
  save_dir = tools::R_user_dir(package = "MultiEWCE", which = "cache")
)
}
\arguments{
\item{file}{File to load.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{save_dir}{Where to store the file locally.}
}
\value{
graph object
}
\description{
Load the HPO as a precomputed graph.
}
\examples{
g <- load_hpo_graph()
}
\keyword{internal}
