% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_celltype.R
\name{map_celltype}
\alias{map_celltype}
\title{Map cell type}
\usage{
map_celltype(
  results,
  input_col = "CellType",
  map = KGExplorer::get_data_package(package = "MSTExplorer", name = "celltype_maps"),
  rm_prefixes = c("Adult", "Fetus", "HESC"),
  by = c("ctd", "author_celltype"),
  add_stage = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{input_col}{Column to use for linking with the \code{map} data.}

\item{map}{1:1 mapping between cell ontology terms and cell type names used
in \code{results}.}

\item{rm_prefixes}{Prefixes to remove from cell type names in
\code{input_col} before performing mapping.}

\item{by}{Columns to merge on.}

\item{add_stage}{Add developmental stage information.}
}
\description{
Map cell types to cell ontology terms.
}
\examples{
results <- load_example_results(multi_dataset = TRUE)
results2 <- map_celltype(results = results)
}
