% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tissue.R
\name{map_tissue}
\alias{map_tissue}
\title{Map cell type}
\usage{
map_tissue(
  results = NULL,
  map = KGExplorer::get_data_package(package = "MSTExplorer", name = "tissue_maps"),
  add_ancestors = 10,
  uberon = KGExplorer::get_ontology(name = "uberon", add_ancestors = add_ancestors),
  return_agg = FALSE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{map}{Many:many mapping between tissue ontology terms and cell type
ontology terms.}

\item{add_ancestors}{Add ancestors for each term.}

\item{uberon}{UBERON ontology object of class \link[simona]{ontology_DAG}.}

\item{return_agg}{Return the aggregated results instead of merging with a
table of association \code{results}.}
}
\description{
Map cell types to cell ontology terms.
}
\examples{
results <- load_example_results()
results2 <- map_tissue(results = results)
}
