% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tissues.R
\name{map_tissues}
\alias{map_tissues}
\title{Add tissues}
\source{
\code{
#### Prepare mapping file #####
URL <- paste0("https://atlas.fredhutch.org/data/bbi/descartes/human_gtex/",
              "downloads/data_summarize_fetus_data/df_cell.RDS")
meta <- readRDS(url(URL))
map <- data.table::data.table(
  stringr::str_split(unique(meta$Organ_cell_lineage),"-",n = 2,
                     simplify = TRUE) |>
    `colnames<-`(c("tissue","celltype"))
)
f <- file.path("~/Desktop/ewce/rare_disease_celltyping/data/",
               "DescartesHuman_celltype_mapping.csv")
               data.table::fwrite(map,f)
piggyback::pb_upload(file = f, repo = "neurogenomics/MSTExplorer")
}
}
\usage{
map_tissues(celltypes = NULL, map = NULL, collapse = NULL)
}
\arguments{
\item{celltypes}{A vector of cell-types to map onto tissues.}

\item{map}{A \link[data.table]{data.table} containing the columns
"celltype" and "tissue".
If \code{NULL}, will use a built-in mapping file that only applies to the
CellTypeDataset DescartesHuman
(see \link[MSTExplorer]{load_example_ctd} for details).}

\item{collapse}{If not \code{NULL},
collapse rows with >1 tissue into one string:
e.g. \code{c("Heart" "Lung")}  --> \code{"Heart;Lung"}}
}
\value{
A list of tissues with the same length
as the input \code{celltypes}.
}
\description{
Annotate the tissues that each cell-type is found in.
}
\examples{
results <- load_example_results()
tissues <- map_tissues(celltypes = results$CellType)
}
