% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results.R
\name{merge_results}
\alias{merge_results}
\title{Merge results}
\usage{
merge_results(save_dir = NULL, res_files = NULL, list_name_column = "hpo_id")
}
\arguments{
\item{save_dir}{Folder to save merged results in.}

\item{res_files}{The output of \link[MSTExplorer]{ewce_para}
(can be either a list of file names, or a nested list of results).}

\item{list_name_column}{The name of the gene_data column
that has the gene list names.}
}
\value{
dataframe and merged results.
}
\description{
Merges individual .rds files from the results output directory into one.
dataframe. Can also handle results lists directly.
Alternatively, can simply provide \code{save_dir} and \emph{.rds}
files will be searched for and imported.
}
\examples{
set.seed(2023)
gene_data <- HPOExplorer::load_phenotype_to_genes()
ctd <- load_example_ctd()
list_names <- unique(gene_data$hpo_id)[seq(3)]
res_files <- ewce_para(ctd = ctd,
                       gene_data = gene_data,
                       list_names = list_names,
                       force_new = TRUE,
                       reps = 10)
all_results <- merge_results(res_files=res_files)
}
