% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiewce_plot.R
\name{multiewce_plot}
\alias{multiewce_plot}
\title{EWCE plot}
\usage{
multiewce_plot(
  results = load_example_results(multi_dataset = TRUE),
  ancestor_names = c("Abnormality of the nervous system",
    "Abnormality of the cardiovascular system", "Abnormality of the immune system",
    "Abnormality of the respiratory system", "Abnormality of the eye",
    "Abnormality of the endocrine system", "Neoplasm"),
  ...
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{...}{
  Arguments passed on to \code{\link[EWCE:ewce_plot]{EWCE::ewce_plot}}
  \describe{
    \item{\code{total_res}}{Results data.frame generated using
\link[EWCE]{bootstrap_enrichment_test} or
\link[EWCE]{ewce_expression_data} functions.
Multiple results tables can be
merged into one results table, as long as the 'list' column is set to
distinguish them.
Multiple testing correction is then applied across all merged results.}
    \item{\code{mtc_method}}{Method to be used for multiple testing correction.
Argument is passed to \link[stats]{p.adjust} (DEFAULT: "bonferroni).}
    \item{\code{q_threshold}}{Corrected significance threshold.}
    \item{\code{ctd}}{CellTypeDataset object.
Should be provided so that the dendrogram can be taken from it
and added to plots.}
    \item{\code{annotLevel}}{An integer indicating which level of \code{ctd} to
analyse (\emph{Default: 1}).}
    \item{\code{heights}}{The relative heights row in the grid. 
Will get repeated to match the dimensions of the grid.
Passed to \link[patchwork]{wrap_plots}.}
    \item{\code{make_dendro}}{Add a dendrogram (requires \code{ctd}).}
    \item{\code{verbose}}{Print messages.}
  }}
}
\value{
A bar chart with dendrogram of EWCE results in each cell type.
}
\description{
A shallow wrapper for the \link[EWCE]{ewce_plot} function.
}
\examples{
results <- load_example_results(multi_dataset=TRUE)
out <- multiewce_plot(results = results)
}
