% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_plot.R
\name{ontology_plot}
\alias{ontology_plot}
\title{Plot graph of phenotypes in the RD EWCE results subsetted by Cell type}
\usage{
ontology_plot(
  cell_type,
  results = load_example_results(),
  color_var = c("fold_change", "q", "p"),
  q_threshold = 5e-04,
  fold_threshold = 1,
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  palette = "Spectral",
  shape = "rect",
  ...
)
}
\arguments{
\item{cell_type}{The cell type of interest to be plotted.
Can be a single string (e.g. \code{"Astrocytes"}) or a character vector
of multiple cell types (e..g. \code{c("Astrocytes","Microglia")}).
Set to \code{NULL} if you wish to include all cell-types that are available
(after \code{q_threshold} and \code{fold_threshold} filtering).
If >1 cell-type remains, results will be aggregated automatically
such that there is only 1 node per phenotype.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{color_var}{Name of the variable (column) in \code{results}
that should be used to color the graph nodes.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{palette}{The colours or colour function that values will be mapped to}

\item{shape}{Character vector of shape names for nodes (or function to set them). Defaults to \code{"circle"}.}

\item{...}{
  Arguments passed on to \code{\link[ontologyPlot:onto_plot]{ontologyPlot::onto_plot}}
  \describe{
    \item{\code{ontology}}{\code{ontology_index} object}
    \item{\code{term_sets}}{List of character vectors of ontological term IDs}
    \item{\code{frequencies}}{Numeric vector of term frequencies named by term IDs}
    \item{\code{terms}}{Character vector of ontological terms}
    \item{\code{edge_attributes}}{List of properties to set for arrows (note, these properties will be used for all arrow).}
    \item{\code{fillcolor}}{Character vector of colours to fill nodes corresponding to \code{terms} with. Alternatively a function to set the colours of the nodes in the graph based on \code{term_sets}.}
    \item{\code{label}}{Character vector of labels (or function to set them).}
    \item{\code{color}}{Character vector of colours for borders of nodes representing \code{terms} (or function to set them).}
    \item{\code{width}}{Numeric vector of widths for nodes (of function to set them).}
    \item{\code{fontsize}}{Numeric vector of font sizes for the text to be placed in the nodes (or function to set them).}
    \item{\code{style}}{Display style for nodes, defaults to \code{"filled"}.}
    \item{\code{fixedsize}}{Character indicating whether nodes should be fixed size, \code{"true"}, or adjusted to fit around the contained text, \code{"false"}.}
  }}
}
\value{
A \pkg{ontologyPlot} plot of the network of phenotypes
in a subset of RD EWCE results.
}
\description{
This subsets the results, using other functions from this script,
by cell type,
q value, and fold change. It then plots a graph/ network of
the phenotypes and
colors the nodes by their fold change or q value
(see the \code{color_var} param).
}
\details{
\emph{Developer notes:}
The ontologyPlot package did not have prebuilt options to create the heatmap,
so the colours are assigned manually to each phenotype in this function.
There must be a more efficient way to do this.
}
\examples{
plt <- ontology_plot(cell_type="Amacrine cells")
}
