% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R
\name{plot_}
\alias{plot_}
\title{Plot functions}
\arguments{
\item{x_var}{Variable to plot on the x-axis.}

\item{y_var}{Variable to plot on the y-axis.}

\item{fill_var}{Variable to fill by.}

\item{facet_var}{Variable to facet by.}

\item{show_plot}{Print the plot to the console.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{y_lab}{y-axis label.}

\item{x_lab}{x-axis label.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{ncol}{Number of facet columns.}

\item{nrow}{Number of facet rows for the plot.}

\item{height}{Height of the saved plot.}

\item{width}{Width of the saved plot.}

\item{heights}{Passed to \link[patchwork]{wrap_plots}.}

\item{subtitle_size}{Size of the plot subtitle.}

\item{phenotype_to_genes}{Phenotype to gene mapping from
\link[HPOExplorer]{load_phenotype_to_genes}.}
}
\value{
R object.
}
\description{
Functions to generate plots.
}
\concept{plot_}
