% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_branches.R
\name{plot_bar_branches}
\alias{plot_bar_branches}
\title{Plot bar plot of enriched phenotypes per celltype per HPO branch}
\usage{
plot_bar_branches(
  results = load_example_results(multi_dataset = TRUE),
  results_full = NULL,
  target_branches = get_target_branches(),
  keep_ancestors = names(target_branches),
  target_celltypes = get_target_celltypes(target_branches = target_branches),
  target_branches_keep = TRUE,
  celltype_col = "cl_name",
  keep_ont_levels = NULL,
  add_test_target_celltypes = TRUE,
  color_map = NULL,
  color_vector = NULL,
  preferred_palettes = "gnuplot",
  legend.position = "none",
  fill_var = "ancestor_name",
  facets = fill_var,
  scales = "free_y",
  q_threshold = NULL,
  lvl = NULL,
  facets_n = "phenotypes_per_ancestor",
  suffix = "phenotypes",
  vbars = NULL,
  ncol = 1,
  cols = NULL,
  y_lab = "Significant phenotypes",
  normalise_by = NULL,
  background_full = TRUE,
  save_path = NULL,
  width = NULL,
  height = NULL
)
}
\description{
Plot bar plot of enriched phenotypes per celltype per HPO branch
}
\examples{
results <- load_example_results(multi_dataset = TRUE)
results <- HPOExplorer::add_ancestor(results,
                                     lvl = 6,
                                     force_new = TRUE)
target_branches <- list("Recurrent bacterial infections"="leukocyte")
out <- plot_bar_branches(results=results,
                         target_branches=target_branches,
                         facets = "hpo_name",
                         legend.position="right",
                         lvl=7,
                         ncol=2,
                         vbars="hepatoblast",
                         facets_n=NULL,
                         q_threshold=0.05,
                         background_full=FALSE)
}
