% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_dendro_facets.R
\name{plot_bar_dendro_facets}
\alias{plot_bar_dendro_facets}
\title{Plot bar branches}
\usage{
plot_bar_dendro_facets(
  results = load_example_results(),
  results_full = NULL,
  target_branches = get_target_branches(),
  keep_ancestors = names(target_branches),
  target_celltypes = get_target_celltypes(target_branches = target_branches),
  target_branches_keep = NULL,
  celltype_col = "cl_name",
  keep_ont_levels = NULL,
  add_test_target_celltypes = TRUE,
  color_map = NULL,
  color_vector = NULL,
  preferred_palettes = "gnuplot",
  legend.position = "none",
  fill_var = "ancestor_name",
  facets = fill_var,
  scales = "free_y",
  q_threshold = NULL,
  lvl = NULL,
  facets_n = "phenotypes_per_ancestor",
  suffix = "phenotypes",
  vlines = NULL,
  ncol = 1,
  cols = NULL,
  y_lab = "Significant phenotypes",
  normalise_by = NULL,
  background_full = TRUE,
  save_path = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{results_full}{A full set of results (not just those selected for
plotting in \code{results}) to be used as a background during proportional
enrichment testing.}

\item{target_branches}{A named list of HPO
branches each matched with CL cell type branches that
correspond to on-target cell types across the two ontologies.}

\item{keep_ancestors}{Only HPO terms that have these ancestors will be kept.}

\item{target_celltypes}{A named list of HPO branches and matched CL ids.}

\item{target_branches_keep}{Whether to plot the target branch ancestors
(\code{TRUE}) or simply its descendants (\code{FALSE}).}

\item{celltype_col}{Name of the cell type column in the \code{results}.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link[HPOExplorer]{add_ont_lvl} for details.}

\item{add_test_target_celltypes}{Using the significant phenotype-cell type
association \code{results}, run proportional enrichment tests to
determine whether each cell type is overrepresented in a given HPO branch
relative to all other HPO branches. Overrepresented cell types will be
denoted by "*" above its bar.}

\item{color_map}{Colour map generated by \link[KGExplorer]{map_colors}
to use for colouring the faceted bars
(i.e. number of significant associations within a given HPO branch).}

\item{color_vector}{A vector of colors to use for colouring the x-axis text
(i.e. cell types).}

\item{preferred_palettes}{Preferred palettes to use for each column.}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{fill_var}{Variable to fill by.}

\item{facets}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{lvl}{HPO ontology to display the final plot as.}

\item{facets_n}{A column to use when displaying the number of unique
phenotypes within each facet.}

\item{suffix}{Suffix for \code{facets_n} text annotation.}

\item{vlines}{Add vertical lines at specific x-axis intercepts
(i.e. cell types).}

\item{ncol}{Number of facet columns.}

\item{cols}{Columns to use for faceting.}

\item{y_lab}{y-axis label.}

\item{normalise_by}{Normalise the bar heights from 0-1 within a given
variable (e.g. "ancestor_name").}

\item{background_full}{Whether to use the full set of results as a
background for proportional enrichment testing.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{width}{Width of the saved plot.}

\item{height}{Height of the saved plot.}
}
\description{
Plot bar plot of enriched phenotypes per cell type per HPO branch.
}
\examples{
hpo <- HPOExplorer::get_hpo()
target_branches <- list("Recurrent bacterial infections"="leukocyte")
lvl <- subset(hpo@elementMetadata,name==names(target_branches)[1])$ontLvl
results <- load_example_results()
results <- HPOExplorer::add_ancestor(results,
                                     lvl = lvl,
                                     force_new = TRUE)
out <- plot_bar_dendro_facets(results=results,
                         target_branches=target_branches,
                         facets = "hpo_name",
                         legend.position="right",
                         lvl=lvl+1,
                         ncol=2,
                         vlines="hepatoblast",
                         facets_n=NULL,
                         q_threshold=0.05,
                         background_full=FALSE)
}
