% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_summary.R
\name{plot_bar_summary}
\alias{plot_bar_summary}
\title{Summary plot}
\usage{
plot_bar_summary(
  results = load_example_results(),
  count_var = "hpo_name",
  group_var = "CellType",
  keywords = NULL,
  q_threshold = 5e-04,
  fold_threshold = 1,
  filters = NULL,
  keep_descendants = NULL,
  hpo = HPOExplorer::get_hpo(),
  option = "magma",
  interactive = TRUE,
  show_plot = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{count_var}{Variable to get counts for per \code{group_var}.}

\item{group_var}{Variable to group counts by.}

\item{keywords}{Keywords supplied to search for phenotypes.
Will be used to generate the plot title.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{option}{A character string indicating the color map option to use.
Eight options are available:
\itemize{
 \item "magma" (or "A")
 \item "inferno" (or "B")
 \item "plasma" (or "C")
 \item "viridis" (or "D")
 \item "cividis" (or "E")
 \item "rocket" (or "F")
 \item "mako" (or "G")
 \item "turbo" (or "H")
}}

\item{interactive}{Make the plot interactive.}

\item{show_plot}{Print the plot after it's been generated.}

\item{verbose}{Print messages.}
}
\value{
ggplot or plotly object
}
\description{
Generate a plot summarising the cell type-phenotype enrichment
results generated by \link[MSTExplorer]{gen_results}.
}
\examples{
keep_descendants <- "Neurodevelopmental delay"
plt_pheno_count <- plot_bar_summary(count_var = "hpo_name",
                                group_var = "CellType",
                                keep_descendants = keep_descendants)
plt_cell_count <- plot_bar_summary(count_var = "CellType",
                                group_var = "hpo_name",
                                keep_descendants = keep_descendants)
}
