% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_severity.R
\name{plot_celltype_severity}
\alias{plot_celltype_severity}
\title{Plot cell type severity}
\usage{
plot_celltype_severity(
  results,
  cl = KGExplorer::filter_ontology(KGExplorer::get_ontology("cl"), keep_descendants =
    "cell"),
  q_threshold = 0.05,
  top_n = 10,
  types = c("bar", "dot")
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{cl}{Cell Ontology (CL) object from
\code{KGExplorer::get_ontology("cl")}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{top_n}{Top and bottom number of cell types to show per annotation
(used in dot plot only).}

\item{types}{Which types of plots of create.}
}
\value{
Named list of ggplot and data.table objects.
}
\description{
Plot the phenotype severity scores (generated by GPT-4) aggregated by
the cell types each phenotype is significantly associated with.
}
\examples{
results <- load_example_results()
out <- plot_celltype_severity(results)
}
