% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_congenital_annotations.R
\name{plot_congenital_annotations}
\alias{plot_congenital_annotations}
\title{Plot congenital annotations}
\usage{
plot_congenital_annotations(
  results,
  gpt_annot = HPOExplorer::gpt_annot_codify(),
  fetal_keywords = c("fetal", "fetus", "primordial", "hESC", "embryonic"),
  celltype_col = "author_celltype",
  remove_annotations = c("varies"),
  save_path = NULL
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{gpt_annot}{A data.table of GPT annotations.}

\item{fetal_keywords}{A character vector of keywords to identify fetal cell types.}

\item{celltype_col}{The column name of the cell type.}

\item{remove_annotations}{A character vector of annotations to remove.}

\item{save_path}{The path to save the plot.}
}
\description{
Test whether there is a difference in proportion if significantly associated
fetal and non-fetal cell types across phenotypes with congenital onset
vs. those without.
}
\examples{
results <- load_example_results()
results2 <- add_ctd(results=results)
}
