% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequency_bar.R
\name{plot_frequency_bar}
\alias{plot_frequency_bar}
\title{Gene and phenotype frequencies bar plot}
\usage{
plot_frequency_bar(
  results = load_example_results(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  allow.cartesian = TRUE,
  show_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{show_plot}{Print the plot after it's been generated.}

\item{verbose}{Print messages.}
}
\value{
ggplot object
}
\description{
Plot the frequency of gene-phenotype and phenotype-disease associations.
}
\examples{
results <- load_example_results()[seq(2500),]
fp_res <- plot_frequency_bar(results=results)
}
