% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequency_histogram.R
\name{plot_frequency_histogram}
\alias{plot_frequency_histogram}
\title{Gene and phenotype frequencies histogram}
\usage{
plot_frequency_histogram(
  results = load_example_results(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  show_plot = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{show_plot}{Print the plot after it's been generated.}

\item{verbose}{Print messages.}
}
\value{
ggplot object
}
\description{
Plot the frequency of gene-phenotype and phenotype-disease associations.
}
\examples{
results <- load_example_results()[seq(500),]
fp_res <- plot_frequency_histogram(results=results)
}
