% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ont_lvl.R
\name{plot_ont_lvl}
\alias{plot_ont_lvl}
\title{Plot ontology levels}
\usage{
plot_ont_lvl(
  results = load_example_results(),
  p2g = HPOExplorer::load_phenotype_to_genes(),
  x_vars = c("genes", "celltypes", "log(abs(fold_change))")
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{p2g}{Phenotype to gene data.}

\item{x_vars}{Variables to plot on the x-axis of each subplot.}
}
\value{
A named list containing the data and the plot.
}
\description{
Generate plots comparing the ontology level of each HPO phenotype and
several other metrics.
}
\examples{
plts <- plot_ont_lvl()
}
