% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report.R
\name{plot_report}
\alias{plot_report}
\title{Report plot}
\usage{
plot_report(
  rep_dt,
  results,
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  annot = HPOExplorer::load_phenotype_to_genes("phenotype.hpoa"),
  remove_cols = c("Rows", "ids"),
  show_plot = TRUE,
  save_plot = tempfile(fileext = "_plot_report.pdf"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{rep_dt}{Report table.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{phenotype_to_genes}{Phenotype to gene mapping from
\link[HPOExplorer]{load_phenotype_to_genes}.}

\item{annot}{HPO annotations.}

\item{remove_cols}{Columns to remove from \code{rep_dt}.}

\item{show_plot}{Show the plot.}

\item{save_plot}{Path to save the plot to.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{filename}}{File name to create on disk.}
    \item{\code{plot}}{Plot to save, defaults to last plot displayed.}
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}
    \item{\code{path}}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width,height,units}}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
  }}
}
\value{
ggplot object
}
\description{
Plot the results of a filtering report generated by
\link[MSTExplorer]{prioritise_targets}.
}
\examples{
results <- load_example_results()
rep_dt <- example_targets$report
gp <- plot_report(rep_dt=rep_dt, results=results)
}
