% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results_distributions.R
\name{plot_results_distributions}
\alias{plot_results_distributions}
\title{Plot results distributions}
\usage{
plot_results_distributions(
  results = load_example_results(),
  x_var = "estimate",
  facets = "CellType",
  binwidth = NULL,
  scales = "fixed",
  log10_transform = TRUE,
  show_plot = TRUE,
  save_path = NULL,
  ...
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{x_var}{Variable to plot on the x-axis.}

\item{facets}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}

\item{binwidth}{The width of the bins. Can be specified as a numeric value
or as a function that calculates width from unscaled x. Here, "unscaled x"
refers to the original x values in the data, before application of any
scale transformation. When specifying a function along with a grouping
structure, the function will be called once per group.
The default is to use the number of bins in \code{bins},
covering the range of the data. You should always override
this value, exploring multiple widths to find the best to illustrate the
stories in your data.

The bin width of a date variable is the number of days in each time; the
bin width of a time variable is the number of seconds.}

\item{scales}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}

\item{log10_transform}{log10 transform \code{x_var}.}

\item{show_plot}{Print the plot to the console.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{...}{
  Arguments passed on to \code{\link[KGExplorer:plot_]{KGExplorer::plot_save}}
  \describe{
    \item{\code{height}}{Height of the heatmap body.}
    \item{\code{width}}{Width of the heatmap body.}
  }}
}
\description{
Plot the distributions of the test statistics in the phenotype-cell type
association results.
}
\examples{
results <- load_example_results()
ct <- unique(results$CellType)[seq(9)]
results <- results[CellType \%in\% ct]
results[,estimate_scaled:=scales::rescale(estimate,c(1,5))]
out <- plot_results_distributions(results, x_var="estimate_scaled")
}
