% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ttd.R
\name{plot_ttd}
\alias{plot_ttd}
\title{Plot TTD}
\usage{
plot_ttd(
  dat_sub,
  base_size = 9,
  label_size = 2,
  as_percent = TRUE,
  baseline = NULL
)
}
\description{
Plot overlap between your targets and TTD targets.
}
\keyword{internal}
