% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_celltypes.R
\name{predict_celltypes}
\alias{predict_celltypes}
\title{Predict cell types}
\usage{
predict_celltypes(
  phenotypes,
  diseases_include = NULL,
  diseases_exclude = NULL,
  genes_include = NULL,
  genes_exclude = NULL,
  gene_weights = list(include = 2, default = 1, exclude = 0),
  results = MSTExplorer::load_example_results(),
  agg_var = c("cl_name"),
  x_var = agg_var[1],
  y_var = "score_sum",
  fill_var = "score_mean",
  max_x_var = 10,
  subtitle_size = 9,
  plot.margin = ggplot2::margin(1, 1, 1, 40),
  show_plot = TRUE,
  save_path = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{phenotypes}{Phenotypes observed in the patient.
Can be a list of HPO phenotype IDs or HPO phenotype names.}

\item{genes_include}{Genes in which the patient is known to have
abnormalities.}

\item{genes_exclude}{Genes in which the patient is known NOT to have
abnormalities.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{diseases}{Diseases that the patient is known to have.
Can be provided as OMIM, Orphanet, or DECIPHER disease IDs.}
}
\value{
\link[data.table]{data.table} of prioritised cell types,
sorted by a "score" that combines:
\itemize{
\item{The phenotype-cell type enrichment p-values ("p").}
\item{The phenotype-cell type enrichment effect size ("fold_change").}
\item{A gene-wise factor that upweights/downweights
included/excluded genes respectively,
multiplied by the evidence score of a phenotype-gene association.
Only applied when \code{genes_include} or \code{genes_exclude} is provided. }
}
}
\description{
Predict the causal cell types underlying a patient's phenotypes given some
varying degree of prior knowledge.
}
\examples{
phenotypes <- c("Generalized neonatal hypotonia",
                "Scrotal hypospadias",
                "Increased circulating progesterone")
# diseases_include <- "OMIM:176270"
genes_include <- c("MAGEL2","HERC2")
genes_exclude <- c("SNORD115-1")
ct <- predict_celltypes(phenotypes = phenotypes,
                        genes_include = genes_include,
                        genes_exclude = genes_exclude)
}
