% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritise_targets_grid.R
\name{prioritise_targets_grid}
\alias{prioritise_targets_grid}
\title{Prioritise targets grid}
\usage{
prioritise_targets_grid(
  top_targets,
  res_class = HPOExplorer::gpt_annot_class(),
  n_per_class = 10,
  keep_severity_class = c("profound", "severe"),
  keep_physical_malformations = NULL,
  species_list = c("Homo sapiens", "Macaca mulatta", "Mus musculus", "Danio rerio",
    "Drosophila melanogaster", "Caenorhabditis elegans"),
  legend.position = "left",
  keep_ont_levels = NULL,
  width = 70,
  widths1 = c(1, 4),
  widths2 = c(1, 0.8),
  show_plot = TRUE,
  ...
)
}
\arguments{
\item{top_targets}{output of \link[MSTExplorer]{prioritise_targets}.}

\item{res_class}{Output of the \link[HPOExplorer]{gpt_annot_class} function.}

\item{n_per_class}{Number of phenotypes per severity class to include.}

\item{keep_severity_class}{Phenotypes to keep based on severity classes.}

\item{keep_physical_malformations}{Phenotypes to keep based on physical
malformation frequency (0=never, 1=rarely, 2=often, 3=always).}

\item{species_list}{Species to include in orthologous genes grid.}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link[HPOExplorer]{add_ont_lvl} for details.}

\item{width}{Width of the saved plot.}

\item{widths1}{Proportional widths of severity annotation heatmap
(left subplot) and the top targets/orthologous genes grid (right subplot).}

\item{widths2}{Proportional widths of the top targets/orthologous genes grid.}

\item{show_plot}{Print the plot to the console.}

\item{...}{
  Arguments passed on to \code{\link[HPOExplorer:plot_top_phenos]{HPOExplorer::plot_top_phenos}}
  \describe{
    \item{\code{keep_descendants}}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}
    \item{\code{annotation_order}}{The order of the annotations to include.}
    \item{\code{split_by_congenital}}{Split the phenotypes by congenital onset
(congenital = always/often, noncongenital = never/rarely).}
    \item{\code{axis.text.x}}{Whether to include x-axis text in top and bottom subplots.}
  }}
}
\value{
A named list with data and patchwork object.
}
\description{
Plot the output of \link[MSTExplorer]{prioritise_targets}
as a grid.
}
\examples{
top_targets <- MSTExplorer::example_targets$top_targets
out <- prioritise_targets_grid(top_targets = top_targets)
}
