% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritise_targets_multiheatmap.R
\name{prioritise_targets_multiheatmap}
\alias{prioritise_targets_multiheatmap}
\title{Prioritise targets: multi-heatmap}
\usage{
prioritise_targets_multiheatmap(
  top_targets,
  prioritise_targets_network_out,
  ctd_list = load_example_ctd(c("ctd_DescartesHuman.rds", "ctd_HumanCellLandscape.rds"),
    multi_dataset = TRUE),
  gene_order = NULL,
  hpo = HPOExplorer::get_hpo(),
  gencc_extra = list(),
  size = 6,
  width = 50,
  heights = NULL,
  show_plot = TRUE
)
}
\arguments{
\item{top_targets}{A results table after it has been annotated with
\link[HPOExplorer]{add_disease} and \link[MSTExplorer]{add_symptom_results}.}

\item{prioritise_targets_network_out}{Output of
\link[MSTExplorer]{prioritise_targets_network}.}

\item{ctd_list}{A named list of CellTypeDataset objects each
created with \link[EWCE]{generate_celltype_data}.}

\item{gene_order}{The order in which to show genes in heatmap x-axis.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{gencc_extra}{Extra rows to add to the diseases GenCC data.}

\item{size}{Plot text size.}

\item{width}{Width of the saved plot.}

\item{heights}{Passed to \link[patchwork]{wrap_plots}.}

\item{show_plot}{Print the plot to the console.}
}
\description{
Prioritise targets: multi-heatmap
}
\examples{
top_targets <- MSTExplorer::example_targets$top_targets[1:10]
prioritise_targets_network_out <- prioritise_targets_network(
  top_targets = top_targets)

out <- prioritise_targets_multiheatmap(
  top_targets = top_targets,
  prioritise_targets_network_out = prioritise_targets_network_out,
  ctd_list=load_example_ctd("ctd_DescartesHuman.rds", multi_dataset = TRUE))
}
