% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priortise_targets.R
\name{priortise_targets}
\alias{priortise_targets}
\title{Prioritise target genes}
\usage{
priortise_targets(
  results,
  ctd,
  annotLevel = 1,
  q_threshold = 0.05,
  fold_threshold = 1,
  keep_celltypes = NULL,
  keep_tiers = c(1, 2),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  hpo = HPOExplorer::get_hpo()
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{ctd}{Cell Type Data List generated using
\link[EWCE]{generate_celltype_data}.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{keep_celltypes}{Cell type to keep.}

\item{keep_tiers}{Tiers from \link[HPOExplorer]{hpo_tiers} to keep.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}
}
\description{
Prioritise target genes based on a procedure:
\enumerate{
\item{Include only results at q<=0.05.}
\item{Include only results at fold_change>=1.}
\item{Include only terminally differentiated cells.}
}
}
\examples{
results <- load_example_results()
ctd <- load_example_ctd()
}
