% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA_load_to_variable.R
\name{RDA_assign_load}
\alias{RDA_assign_load}
\title{Load RDA file and assign to specific variable
Can be useful, but sort of pointless now as we have decided to only use .rds}
\usage{
RDA_assign_load(file)
}
\arguments{
\item{file}{file path to rda \<string\>}
}
\value{
the data contained in the rda file
}
\description{
Load RDA file and assign to specific variable
Can be useful, but sort of pointless now as we have decided to only use .rds
}
\examples{

\dontrun{
files = "results.rds"
RDA_assign_load(file)}

}
