% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_congenital_enrichment.R
\name{run_congenital_enrichment}
\alias{run_congenital_enrichment}
\title{Run congenital enrichment analyses}
\usage{
run_congenital_enrichment(
  results,
  hpo = HPOExplorer::get_hpo(),
  cl = get_cl(),
  gpt_annot = HPOExplorer::gpt_annot_codify(),
  fetal_keywords = c("fetal", "fetus", "primordial", "hESC", "embryonic"),
  celltype_col = "author_celltype",
  top_n_phenotypes = 10,
  top_n_hpo = 50,
  top_n_cl = top_n_hpo,
  q_threshold = 0.05,
  prune = TRUE,
  ...
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{cl}{Cell Ontology (CL) object from
\code{KGExplorer::get_ontology("cl")}.}

\item{gpt_annot}{A data.table of GPT annotations.}

\item{fetal_keywords}{A character vector of keywords to identify fetal cell types.}

\item{celltype_col}{The column name of the cell type.}

\item{top_n_phenotypes}{Number of top and bottom phenotypes to return.}

\item{top_n_hpo}{Number of top and bottom phenotypes
to run enrichment analyses on.}

\item{top_n_cl}{Number of top and bottom cell types
to run enrichment analyses on.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{prune}{Prune redundant ancestral terms from the enrichment results
using \link[KGExplorer]{prune_ancestors}.}

\item{...}{
  Arguments passed on to \code{\link[simona:dag_enrich_on_offsprings]{simona::dag_enrich_on_offsprings}}
  \describe{
    \item{\code{dag}}{An \code{ontology_DAG} object.}
    \item{\code{terms}}{A vector of term names.}
    \item{\code{min_hits}}{Minimal number of terms in an offspring set.}
    \item{\code{min_offspring}}{Minimal size of the offspring set.}
  }}
}
\value{
Named list of enrichment results.
}
\description{
First, this function computes the mean difference in association p-values
between a given phenotype and the congenital and non-congenital versions
of a cell type (the \code{fetal_nonfetal_pdiff} column).
It then sorts the results from the largest to the smallest difference.
Here, a positive difference indicates that the phenotype is more associated
with the fetal/embryonic version of the cell type, while a negative difference
indicates that the phenotype is more associated with the adult version.
Next, it runs enrichment analyses for the top and bottom N phenotypes
using the \link[simona]{dag_enrich_on_offsprings} function.
Finally, it run enrichment analyses for the top and bottom N cell types.
}
\examples{
results <- load_example_results()[ctd=="HumanCellLandscape"]
out <- run_congenital_enrichment(results=results)
}
