% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_prop_tests.R
\name{run_prop_tests}
\alias{run_prop_tests}
\title{Run proportional enrichment tests}
\usage{
run_prop_tests(
  results,
  branch_col = "ancestor_name",
  celltype_col = "CellType",
  func = list(rstatix::fisher_test, rstatix::prop_test)[[1]],
  alternative = "greater",
  cores = NULL,
  ...
)
}
\description{
Run a series of proportional enrichment tests on the results of a
phenotype to cell type association test.
}
\examples{
results <- load_example_results()
results <- HPOExplorer::add_ancestor(results)
}
\keyword{internal}
