% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise_genes.R
\name{standardise_genes}
\alias{standardise_genes}
\title{Standardise genes}
\usage{
standardise_genes(
  dat,
  gene_col = "gene_symbol",
  fill_na = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{data.table}

\item{gene_col}{character column name.}

\item{fill_na}{logical. Fill NAs with original gene symbol.}

\item{verbose}{logical. Print messages.}

\item{...}{
  Arguments passed on to \code{\link[orthogene:map_genes]{orthogene::map_genes}}
  \describe{
    \item{\code{genes}}{Gene list.}
    \item{\code{species}}{Species to map against.}
    \item{\code{target}}{target namespace.}
    \item{\code{mthreshold}}{maximum number of results per initial alias to show. Shows all by default.}
    \item{\code{drop_na}}{Drop all genes without mappings.
Sets \code{gprofiler2::gconvert(filter_na=)} as well
an additional round of more comprehensive \code{NA} filtering
by \pkg{orthogene}.}
    \item{\code{numeric_ns}}{namespace to use for fully numeric IDs (\href{https://biit.cs.ut.ee/gprofiler/page/namespaces-list}{list of available namespaces}).}
    \item{\code{run_map_species}}{Standardise \code{species} names with 
\link[orthogene]{map_species} first (Default: \code{TRUE}).}
  }}
}
\value{
data.table
}
\description{
Standardise gene symbols to HGNC.
}
\examples{
\dontrun{
dat <- data.table(gene_symbol = c("BRCA1","BRCA2","BRCA3"))
dat2 <- standardise_genes(dat)
}
}
