% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_phenos.R
\name{subset_phenos}
\alias{subset_phenos}
\title{Subset RD EWCE results}
\usage{
subset_phenos(
  cell_type,
  results = load_example_results(),
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  verbose = TRUE
)
}
\arguments{
\item{cell_type}{The cell type of interest to be plotted.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{verbose}{Print messages.}
}
\value{
A data frame of results taken from the main data frame of results.
}
\description{
This subsets  the Rare disease EWCE results by cell type,
q threshold and fold change.
}
\examples{
phenos <- subset_phenos(cell_type = "Amacrine cells")
}
