% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_results.R
\name{subset_results}
\alias{subset_results}
\title{Subset EWCE results}
\usage{
subset_results(
  cell_type,
  results = load_example_results(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  hpo = HPOExplorer::get_hpo(),
  verbose = TRUE
)
}
\arguments{
\item{cell_type}{The cell type of interest to be plotted.
Can be a single string (e.g. \code{"Astrocytes"}) or a character vector
of multiple cell types (e..g. \code{c("Astrocytes","Microglia")}).
Set to \code{NULL} if you wish to include all cell-types that are available
(after \code{q_threshold} and \code{fold_threshold} filtering).
If >1 cell-type remains, results will be aggregated automatically
such that there is only 1 node per phenotype.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{verbose}{Print messages.}
}
\value{
A data frame of the selected subset of RD EWCE results
with HPO ID column added.
}
\description{
Subset RD EWCE results data by cell type, fold change and q value.
}
\examples{
signif_cell_data <- subset_results(cell_type="Cardiomyocytes")
}
