% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise results}
\usage{
summarise_results(
  results,
  group_var = "ctd",
  add_merged = TRUE,
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  digits = 3,
  save_path = tempfile("summarise_results.csv")
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{group_var}{Variable to segregate results by.}

\item{add_merged}{Add a merged summary across all groups.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{digits}{a positive integer indicating how many significant digits
    are to be used for
    numeric and complex \code{x}.  The default, \code{NULL}, uses
    \code{\link[base]{getOption}("digits")}.  This is a suggestion: enough decimal
    places will be used so that the smallest (in magnitude) number has
    this many significant digits, and also to satisfy \code{nsmall}.
    (For more, notably the interpretation for complex numbers see
    \code{\link[base]{signif}}.)}

\item{save_path}{Path to save the summary to as a CSV.}
}
\description{
Summarise results from \pkg{MSTExplorer}.
}
\examples{
results <- load_example_results()
summary_split <- summarise_results(results, group_var="ctd")
summary_merged <- summarise_results(results, group_var=NULL)
}
