% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Summary plot}
\usage{
summary_plot(
  results = load_example_results(),
  count_var = "hpo_name",
  group_var = "CellType",
  keywords = NULL,
  q_threshold = 5e-04,
  fold_threshold = 1,
  cell_type = NULL,
  ancestor = NULL,
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  option = "magma",
  interactive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{count_var}{Variable to get counts for per \code{group_var}.}

\item{group_var}{Variable to group counts by.}

\item{keywords}{Keywords supplied to search for phenotypes.
Will be used to generate the plot title..}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{cell_type}{The cell type of interest to be plotted.
Can be a single string (e.g. \code{"Astrocytes"}) or a character vector
of multiple cell types (e..g. \code{c("Astrocytes","Microglia")}).
Set to \code{NULL} if you wish to include all cell-types that are available
(after \code{q_threshold} and \code{fold_threshold} filtering).
If >1 cell-type remains, results will be aggregated automatically
such that there is only 1 node per phenotype.}

\item{ancestor}{The ancestor to get all descendants of. If \code{NULL},
returns the entirely ontology.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{option}{A character string indicating the color map option to use.
Eight options are available:
\itemize{
 \item "magma" (or "A")
 \item "inferno" (or "B")
 \item "plasma" (or "C")
 \item "viridis" (or "D")
 \item "cividis" (or "E")
 \item "rocket" (or "F")
 \item "mako" (or "G")
 \item "turbo" (or "H")
}}

\item{interactive}{Make the plot interactive with \link[plotly]{ggplotly}.}

\item{verbose}{Print messages.}
}
\value{
ggplot or plotly object
}
\description{
Generate a plot summarising the cell type-phenotype enrichment
results generated by \link[MultiEWCE]{gen_results}.
}
\examples{
ancestor <- "Neurodevelopmental delay"
plt_pheno_count <- summary_plot(count_var = "hpo_name",
                                group_var = "CellType",
                                ancestor = ancestor)
plt_cell_count <- summary_plot(count_var = "CellType",
                                group_var = "hpo_name",
                                ancestor = ancestor)
}
