% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminal_celltypes.R
\name{terminal_celltypes}
\alias{terminal_celltypes}
\title{Terminal cell types}
\usage{
terminal_celltypes()
}
\value{
data.table of cell types.
}
\description{
A manually-curated list of terminally differentiated cell types
along with supporting references from the literature.
}
\examples{
celltypes <- terminal_celltypes()
}
