% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_target_celltypes.R
\name{test_target_celltypes}
\alias{test_target_celltypes}
\title{Test target cell types}
\usage{
test_target_celltypes(
  results = load_example_results(),
  target_celltypes = get_target_celltypes(),
  celltype_col = "cl_id",
  tests = c("within_branches", "across_branches", "across_branches_per_celltype"),
  within = "hpo_id",
  ancestor_var = "ancestor_name",
  q_threshold = 0.05,
  cores = 1
)
}
\arguments{
\item{within}{(optional) within-subjects factor variables}

\item{cores}{The number of cores to run in parallel (e.g. 8) \code{int}.}
}
\description{
Test the following hypotheses across \link{gen_results} enrichment results.
\itemize{
\item{"within_branches"}{
Are on-target cell types enriched for significant results?}
\item{"across_branches"}{
Are on-target cell types enriched for significant results across branches?}
\item{"across_branches_per_celltype"}{
Are on-target cell types enriched for significant results across
branches per cell type?}
}
}
\examples{
res <- test_target_celltypes(tests="within_branches")
}
