% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tissue_maps}
\alias{tissue_maps}
\title{Anatomy maps}
\format{
data.table
}
\source{
\code{
## Download standardised Seurat objects from CellXGene.
obj_list <- list()
obj_list[["DescartesHuman"]] <- readRDS("~/Downloads/db4d63ab-0ca3-4fa5-b0cf-34955227912d.rds")
obj_list[["HumanCellLandscape"]] <- readRDS("~/Downloads/1b7484e3-83a0-47fe-847e-54d811a2adae.rds")
cols <- c("tissue_ontology_term_id",
          "tissue",
          # "tissue_original",
          "cell_type_ontology_term_id",
          "cell_type")
tissue_maps <- lapply(obj_list, function(obj){
  select_cols <- cols[cols %in% colnames(obj@meta.data)]
  d <- data.table::as.data.table(obj@meta.data[,select_cols])
  d[,cl_count:=.N, by=c("cell_type_ontology_term_id","cell_type")]
  unique(d)
}) |> data.table::rbindlist(idcol = "ctd", fill=TRUE)
data.table::setnames(tissue_maps,
                     c("tissue_ontology_term_id","tissue",
                       "cell_type_ontology_term_id","cell_type"),
                     c("uberon_id","uberon_name",
                       "cl_id","cl_name"))
usethis::use_data(tissue_maps, overwrite = TRUE)
}
}
\usage{
data("tissue_maps")
}
\description{
Mappings from cell-type names to anatomy ontology terms in the
HumanCellLandscape and DescartesHuman scRNA-seq datasets.
}
\keyword{datasets}
