% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttd_check.R
\name{ttd_check}
\alias{ttd_check}
\title{Gene therapy targets}
\usage{
ttd_check(
  top_targets,
  drug_types = NULL,
  failed_status = c("Terminated", "Withdrawn from market", NA),
  keep_status = NULL,
  remove_status = c(NA),
  allow.cartesian = FALSE,
  show_plot = TRUE,
  save_path = NULL,
  height = NULL,
  width = NULL
)
}
\arguments{
\item{top_targets}{Top targets generated by
\link[MSTExplorer]{prioritise_targets}.}

\item{drug_types}{Filter results by drug type.}

\item{keep_status}{Filter results by drug approval status.}

\item{show_plot}{Show the plot.}
}
\description{
Identify the overlap between your prioritised list of gene therapy targets
and currently existing gene therapy targets that are currently on the market
or are in clinical trials. Uses data from the
\href{https://db.idrblab.net/ttd/full-data-download}{
Therapuetic Target Database}.
}
\examples{
top_targets <- MSTExplorer::example_targets$top_targets
res <- ttd_check(top_targets=top_targets)
}
