% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_associations_correlate_ctd.R
\name{validate_associations_correlate_ctd}
\alias{validate_associations_correlate_ctd}
\title{Validate associations: correlate CellTypeDataset}
\usage{
validate_associations_correlate_ctd(
  results = load_example_results(),
  filters = NULL,
  hpo_agg_lvl = 3,
  hpo = HPOExplorer::get_hpo(),
  group_var = "ctd",
  celltype_var = "cl_name",
  q_threshold = 0.05,
  downsample = NULL,
  add_density = TRUE,
  marginal = FALSE,
  rasterize_points = TRUE,
  dpi = 100,
  point.args = list(alpha = 0.5),
  stats_idx = c(1, 3, 4, 7),
  ...
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{hpo_agg_lvl}{Aggregate the data to a specific HPO ancestor level
during plotting to reduce figure size.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{group_var}{A character string specifying the column in \code{results}
that contains the group variable to compare across.}

\item{celltype_var}{A character string specifying the column in \code{results}
that contains the cell type variable to compare across.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{downsample}{Downsample the data to this many points when plotting.}

\item{add_density}{Add 2D density plots to the scatter plots.}

\item{marginal}{Decides whether marginal distributions will be plotted on
axes using \code{{ggside}} functions. The default is \code{TRUE}. The package
\code{{ggside}} must already be installed by the user.}

\item{rasterize_points}{Whether to rasterize the points in the scatter plots.}

\item{dpi}{Resolution of image after rasterization.}

\item{point.args}{A list of additional aesthetic arguments to be passed to
the \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{stats_idx}{Indices of statistics to keep in subtitle}

\item{...}{Additional arguments passed to \code{plot_density_cor}.}
}
\description{
Validate phenotype-cell type associations can making comparisons within and
between results from different CellTypeDatasets.
}
\examples{
results <- load_example_results()[,.SD[seq(10000)],by="ctd"]
#### Across CTD ####
out1 <- validate_associations_correlate_ctd(results=results,
                                            group_var="ctd")
#### Within CTD: across developmental stages ####
filters <- list(ctd=c("HumanCellLandscape"), stage=c("Fetus","Adult"))
out2 <- validate_associations_correlate_ctd(results=results,
                                            filters=filters,
                                            group_var="stage")
}
