% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_associations_mkg.R
\name{validate_associations_mkg}
\alias{validate_associations_mkg}
\title{Validate associations with the Monarch Knowledge Graph}
\usage{
validate_associations_mkg(
  results = load_example_results(),
  kg = get_data("monarch_kg_cells.csv"),
  q_threshold = 0.05,
  cl = get_cl(),
  alt_ids = c(`CL:0000111` = "CL:2000032"),
  metric = c("dist_nca.min_adj", "dist_lca.min_adj", "dist_nca.min", "dist_lca.min")[1]
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{kg}{A data.table with phenotype-cell type relationships
("from" and "to" columns, respectively) gathered from
the Monarch Knowledge Graph.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{cl}{Cell Ontology object.}

\item{alt_ids}{Alternative IDs to map CL IDs into.}

\item{metric}{Ontological distance metric to use for plotting.}
}
\description{
This function validates the associations between phenotypes and cell types
with the Monarch Knowledge Graph (MKG). It computes the number of phenotypes
and cell types that are present in the MKG and the number of associations
that are present in the MKG.
}
\examples{
results <- load_example_results()
kg <- validate_associations_mkg(result=results)
}
