% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_symptom_results.R
\name{add_symptom_results}
\alias{add_symptom_results}
\title{Add symptom results}
\usage{
add_symptom_results(
  results = load_example_results(),
  q_threshold = 0.05,
  effect_threshold = NULL,
  celltype_col = "CellType",
  ctd_list = load_example_ctd(file = paste0("ctd_", unique(results$ctd), ".rds"),
    multi_dataset = TRUE),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  annotLevels = map_ctd_levels(results),
  keep_quantiles = seq(30, 40),
  top_n = NULL,
  proportion_driver_genes_symptom_threshold = 0.25,
  drop_subthreshold = FALSE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{effect_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{celltype_col}{Cell type column name in \code{results}.}

\item{ctd_list}{A named list of CellTypeDataset objects each
created with \link[EWCE]{generate_celltype_data}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{annotLevels}{The annotation level to use within each CTD in
\code{ctd_list}.}

\item{keep_quantiles}{Quantiles to keep in each CellTypeDataset of the
\code{ctd_list}.}

\item{top_n}{Top N genes to keep when grouping by \code{group_vars}.}

\item{proportion_driver_genes_symptom_threshold}{The minimum proportion of
overlap between symptom genes (genes annotated to a phenotype
via a specific disease) and the driver genes
(genes driving a signficant phenotype-cell type association).}

\item{drop_subthreshold}{Drop rows that don't meet the
\code{proportion_driver_genes_symptom_threshold} criterion.}
}
\description{
Add symptom results to the results data.table.
}
\examples{
results <- load_example_results()[seq(5000)]
results <- add_symptom_results(results)
}
