% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_results.R
\name{agg_results}
\alias{agg_results}
\title{Aggregate results}
\usage{
agg_results(
  phenos,
  count_var = "Phenotype",
  group_var = "CellType",
  sep = "; ",
  verbose = TRUE
)
}
\arguments{
\item{phenos}{dataframe of phenotypes and values / parameters.}

\item{count_var}{Variable to get counts for per \code{group_var}.}

\item{group_var}{Variable to group counts by.}

\item{sep}{Separator for collapsed character columns.}

\item{verbose}{Print messages.}
}
\value{
Aggregated \link[data.table]{data.table}
}
\description{
Aggregate enrichment results from generated by \link[MultiEWCE]{gen_results}.
Counts unique entries in \code{count_var} while grouping by \code{group_var}.
}
\examples{
phenos <- subset_results(cell_type = "Microglia")
agg_res <- agg_results(phenos = phenos)
}
