% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_heatmap.R
\name{correlation_heatmap}
\alias{correlation_heatmap}
\title{Plot heatmap}
\usage{
correlation_heatmap(
  top_targets,
  row_side_vars = c("ancestor_name", "ontLvl"),
  col_side_vars = c("n_celltypes", "n_genes"),
  row_labels = "hpo_name",
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  col = pals::gnuplot(),
  show_plot = TRUE,
  save_path = tempfile(fileext = "correlation_heatmap.pdf"),
  height = 10,
  width = height * 1.3,
  fontsize = 7,
  seed = 2023,
  verbose = TRUE
)
}
\arguments{
\item{top_targets}{\link[data.table]{data.table} of prioritised targets
generated by \link[MSTExplorer]{prioritise_targets}.}

\item{row_side_vars}{Variables to include in row-side
metadata annotations.}

\item{col_side_vars}{Variables to include in column-side
metadata annotations.}

\item{row_labels}{Optional row labels which are put as row names in the heatmap.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{show_plot}{Show the plot.}

\item{save_path}{Path to save interactive plot to
as a self-contained HTML file.}

\item{height}{Height of the heatmap body.}

\item{width}{Width of the heatmap body.}

\item{fontsize}{Axis labels font size.}

\item{seed}{Set the seed for reproducible clustering.}

\item{verbose}{Print messages.}
}
\value{
Plot
}
\description{
Plot a phenotype x phenotype correlation matrix based on genetic overlap.
}
\examples{
top_targets <- MSTExplorer::example_targets$top_targets[seq(100),]
hm <- correlation_heatmap(top_targets = top_targets)
}
