% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewce_para.R
\name{ewce_para}
\alias{ewce_para}
\title{EWCE parallel}
\usage{
ewce_para(
  ctd,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  list_names = unique(gene_data[[list_name_column]]),
  bg = unique(gene_data[[gene_column]]),
  reps = 100,
  annotLevel = 1,
  genelistSpecies = "human",
  sctSpecies = "human",
  save_dir_tmp = tempdir(),
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{ctd}{Cell Type Data List generated using
\link[EWCE]{generate_celltype_data}.}

\item{gene_data}{data frame of gene list names and genes
(see \link[HPOExplorer]{get_gene_lists}).}

\item{list_name_column}{The name of the gene_data column
that has the gene list names.}

\item{gene_column}{The name of the gene_data column that contains the genes.}

\item{list_names}{character vector of gene list names.}

\item{bg}{List of gene symbols containing the background gene list
(including hit genes). If \code{bg=NULL},
 an appropriate gene background will be created automatically.}

\item{reps}{Number of random gene lists to generate (\emph{Default: 100},
but should be >=10,000 for publication-quality results).}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{genelistSpecies}{Species that \code{hits} genes came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{sctSpecies}{Species that \code{sct_data} came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{save_dir_tmp}{Folder to save intermediate results files to
(one file per gene list). Set to \code{NULL} to skip saving temporary files.}

\item{cores}{The number of cores to run in parallel (e.g. 8) \code{int}.}

\item{verbose}{Print messages.}
}
\value{
Paths to saved results at "(save_dir)/(list_name).rds"
(when \code{save_dir!=NULL}), or a nested list of results
(when \code{save_dir==NULL}).
}
\description{
Runs EWCE in parallel on multiple gene lists.
}
\examples{
gene_data <- HPOExplorer::load_phenotype_to_genes()
ctd <- MultiEWCE::load_example_ctd()
list_names <- unique(gene_data$Phenotype)[seq_len(3)]
res_files <- ewce_para(ctd = ctd,
                       gene_data = gene_data,
                       list_names = list_names,
                       reps = 10)
}
