% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewce_para.R
\name{ewce_para}
\alias{ewce_para}
\title{EWCE parallel}
\usage{
ewce_para(
  list_names,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  results_directory,
  ctd_file,
  background_genes,
  bootstrap_reps,
  annotation_Level,
  genes_Species,
  ctd_Species,
  cores
)
}
\arguments{
\item{list_names}{character vector of gene list names}

\item{gene_data}{data frame of gene list names and genes (see ?get_gene_list)}

\item{list_name_column}{The name of the gene_data column that has the gene list names}

\item{gene_column}{The name of the gene_data column that contains the genes}

\item{results_directory}{The desired output filepath for results to be saved}

\item{ctd_file}{The cell type data object for EWCE analysis (see EWCE docs)}

\item{background_genes}{The background geneset for EWCE analysis (see EWCE docs)}

\item{bootstrap_reps}{The number of bootstrap reps <int> (e.g. 100000)}

\item{annotation_Level}{The level of desired cell resolution from the CTD}

\item{genes_Species}{The species of gene lists <string> "human" or "mouse"}

\item{ctd_Species}{"human" or "mouse" <string>}

\item{cores}{The number of cores to run in parallel (e.g. 8) <int>}
}
\value{
True if analysis was sucessful,
saves results at "<results_directory>/<list_name>.rds"
}
\description{
Runs EWCE in parallel on multiple gene lists.
}
