% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewce_plot.R
\name{ewce_plot}
\alias{ewce_plot}
\title{EWCE plot}
\usage{
ewce_plot(total_res, mtc_method = "bonferroni", ctd = NULL, ...)
}
\arguments{
\item{total_res}{Results data.frame generated using
\link[EWCE]{bootstrap_enrichment_test} or
\link[EWCE]{ewce_expression_data} functions.
Multiple results tables can be
merged into one results table, as long as the 'list' column is set to
distinguish them.
Multiple testing correction is then applied across all merged results.}

\item{mtc_method}{Method to be used for multiple testing correction.
Argument is passed to \link[stats]{p.adjust} (DEFAULT: "bonferroni).}

\item{ctd}{CellTypeDataset object.
Should be provided so that the dendrogram can be taken from it
and added to plots.}

\item{...}{
  Arguments passed on to \code{\link[EWCE:ewce_plot]{EWCE::ewce_plot}}
  \describe{
    \item{\code{q_threshold}}{Corrected significance threshold.}
    \item{\code{annotLevel}}{An integer indicating which level of \code{ctd} to
analyse (\emph{Default: 1}).}
    \item{\code{heights}}{The relative heights row in the grid. 
Will get repeated to match the dimensions of the grid.
Passed to \link[patchwork]{wrap_plots}.}
    \item{\code{make_dendro}}{Add a dendrogram (requires \code{ctd}).}
    \item{\code{verbose}}{Print messages.}
  }}
}
\value{
A bar chart with dendrogram of EWCE results in each cell type.
}
\description{
A shallow wrapper for the \link[EWCE]{ewce_plot} function.
}
\examples{
full_results <- EWCE::example_bootstrap_results()
plt <- MultiEWCE::ewce_plot(total_res = full_results$results)
}
