% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_results.R
\name{gen_results}
\alias{gen_results}
\title{Generate results}
\usage{
gen_results(
  ctd,
  gene_data,
  list_name_column = "hpo_id",
  gene_column = "gene_symbol",
  list_names = unique(gene_data[[list_name_column]]),
  reps = 100,
  annotLevel = 1,
  genelistSpecies = "human",
  sctSpecies = "human",
  force_new = FALSE,
  bg = get_bg(species1 = genelistSpecies, species2 = sctSpecies, overwrite = force_new),
  min_genes = 4,
  cores = 1,
  parallel_boot = FALSE,
  save_dir_tmp = NULL,
  save_dir = tempdir(),
  verbose = 1,
  ...
)
}
\arguments{
\item{ctd}{CellTypeDataset generated using
\link[EWCE]{generate_celltype_data}.}

\item{gene_data}{data frame of gene list names and genes
(see \link[HPOExplorer]{get_gene_lists}).}

\item{list_name_column}{The name of the gene_data column
that has the gene list names.}

\item{gene_column}{The name of the gene_data column that contains the genes.}

\item{list_names}{character vector of gene list names.}

\item{reps}{Number of random gene lists to generate (\emph{Default: 100},
but should be >=10,000 for publication-quality results).}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{genelistSpecies}{Species that \code{hits} genes came from
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{sctSpecies}{Species that \code{sct_data} is currently formatted as
(no longer limited to just "mouse" and "human").
See \link[EWCE]{list_species} for all available species.}

\item{force_new}{Overwrite previous results
in the \code{save_dir_tmp}.}

\item{bg}{List of gene symbols containing the background gene list
(including hit genes). If \code{bg=NULL},
 an appropriate gene background will be created automatically.}

\item{min_genes}{Minimum number of genes per list (default: 4)}

\item{cores}{The number of cores to run in parallel (e.g. 8) \code{int}.}

\item{parallel_boot}{Parallelise at the level of bootstrap iterations,
rather than across gene lists.}

\item{save_dir_tmp}{Folder to save intermediate results files to
(one file per gene list). Set to \code{NULL} to skip saving temporary files.}

\item{save_dir}{Folder to save merged results in.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[EWCE:bootstrap_enrichment_test]{EWCE::bootstrap_enrichment_test}}
  \describe{
    \item{\code{sct_data}}{List generated using \link[EWCE]{generate_celltype_data}.}
    \item{\code{hits}}{List of gene symbols containing the target gene list.
Will automatically be converted to human gene symbols
if \code{geneSizeControl=TRUE}.}
    \item{\code{sctSpecies_origin}}{Species that the \code{sct_data} 
originally came from, regardless of its current gene format 
(e.g. it was previously converted from mouse to human gene orthologs).
 This is used for computing an appropriate backgrund.}
    \item{\code{output_species}}{Species to convert \code{sct_data} and \code{hits} to
(Default: "human").
See \link[EWCE]{list_species} for all available species.}
    \item{\code{method}}{R package to use for gene mapping:
\itemize{
\item{\code{"gprofiler"} : Slower but more species and genes.}
\item{\code{"homologene"} : Faster but fewer species and genes.}
\item{\code{"babelgene"} : Faster but fewer species and genes.
Also gives consensus scores for each gene mapping based on a
 several different data sources.}
}}
    \item{\code{no_cores}}{Number of cores to parallelise
bootstrapping \code{reps} over.}
    \item{\code{geneSizeControl}}{Whether you want to control for
GC content and transcript length. Recommended if the gene list originates
from genetic studies (\emph{Default: FALSE}).
If set to \code{TRUE}, then \code{hits} must be from humans.}
    \item{\code{controlledCT}}{[Optional] If not NULL, and instead is the name of a
cell type, then the bootstrapping controls for expression within that
cell type.}
    \item{\code{mtc_method}}{Multiple-testing correction method
(passed to \link[stats]{p.adjust}).}
    \item{\code{sort_results}}{Sort enrichment results from
smallest to largest p-values.}
    \item{\code{localHub}}{If working offline, add argument localHub=TRUE to work 
with a local, non-updated hub; It will only have resources available that
have previously been downloaded. If offline, Please also see BiocManager
vignette section on offline use to ensure proper functionality.}
  }}
}
\value{
All results as a dataframe.
}
\description{
Generates EWCE results on multiple gene lists in parallel by calling
\code{ewce_para}. It allows you to stop the analysis and then continue later
from where you left off as it checks the results output directory for
finished gene lists and removes them from the input.
It also excludes gene lists with
less than 4 unique genes (which cause errors in EWCE analysis).
}
\details{
The gene_data should be a data frame that contains a column of gene
list names (e.g. the column may be called "hpo_name"), and a column of
genes (e.g. "gene_symbol"). For example:\tabular{ll}{
   hpo_name \tab gene_symbol \cr
   "Abnormal heart" \tab gene X \cr
   "Abnormal heart" \tab gene Y \cr
   "Poor vision" \tab gene Z \cr
   "Poor vision" \tab gene Y \cr
   etc... \tab  \cr
}


For more information on this see docs for get_gene_list
(\link[HPOExplorer]{get_gene_lists}).
}
\examples{
gene_data <- HPOExplorer::load_phenotype_to_genes()
list_names <- unique(gene_data$hpo_id)[seq(5)]
ctd <- load_example_ctd()
all_results <- gen_results(ctd = ctd,
                           gene_data = gene_data,
                           list_names = list_names,
                           reps = 10)
}
