% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_results.R
\name{gen_results}
\alias{gen_results}
\title{Generate results}
\usage{
gen_results(
  ctd,
  gene_data,
  list_names,
  background_genes,
  list_name_column = "Phenotype",
  gene_column = "Gene",
  results_dir = "results",
  overwrite_past_analysis = FALSE,
  reps = 10,
  annotLevel = 1,
  genelistSpecies = "human",
  sctSpecies = "human",
  cores = 1,
  MergeResults = TRUE
)
}
\arguments{
\item{ctd}{The Cell type data file for EWCE analysis (see EWCE docs)}

\item{gene_data}{The dataframe containing gene list names and associated genes
(see docs for get_gene_list for more info).}

\item{list_names}{The names of each gene list (e.g. "Abnormality of nervous system
may be the name of a phenotype assocated gene list)}

\item{background_genes}{A character vector of background genes (see EWCE docs)}

\item{list_name_column}{The name of the column in gene_data that contains the
gene list names, (e.g. the column may be called "Phenotype" if dealing with
phenotype associated gene lists)}

\item{gene_column}{The name of the column containing genes in the gene_data
dataframe. Typically this column is called "Gene"}

\item{results_dir}{the desired direcory to save results (e.g. "results")}

\item{overwrite_past_analysis}{overwrite previous results in the results dir <bool>}

\item{reps}{The number of bootstrap reps for EWCE (see ewce docs) <int>}

\item{annotLevel}{The level of cell specificity to select from the CTD,
See EWCE docs <int>}

\item{genelistSpecies}{The species ("human"/"mouse") of the gene lists <string>}

\item{sctSpecies}{The species ("human"/"mouse") of the CTD data}

\item{cores}{The number of cores to run in parallel <int>}

\item{MergeResults}{Return a single dataframe of all results merged <bool>}
}
\value{
If MergeResults == TRUE, it will return all results as a datframe. If
FALSE nothing will be returned, but the individual results will still be saved
in the results directory.
}
\description{
Generates EWCE results on multiple gene lists in parallel by calling
\code{ewce_para}. It allows you to stop the analysis and then continue later
from where you left off as it checks the results output directory for finished
gene lists and removes them from the input. It also excludes gene lists with
less than 4 unique genes (which cause errors in ewce analysis).
}
\details{
The gene_data should be a data frame that contains a column of gene
list names (e.g. the column may be called "Phenotype"), and a column of
genes (e.g. "Gene"). For example:\tabular{ll}{
   Phenotype \tab Gene \cr
   "Abnormal heart" \tab gene X \cr
   "Abnormal heart" \tab gene Y \cr
   "Poor vision" \tab gene Z \cr
   "Poor vision" \tab gene Y \cr
   etc... \tab  \cr
}


For more information on this see docs for get_gene_list (\code{?get_gene_list})

If \code{MergeResutls == TURE}, the function will return a dataframe of all results.
No multiple testing corrections are applied to this so it is recommended that
they are done after, for example:
\code{all_results$q <- stats::p.adjust(all_results$p, method = "BH")}
}
