% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_list.R
\name{get_gene_list}
\alias{get_gene_list}
\title{get_gene_list}
\usage{
get_gene_list(
  list_name,
  gene_data,
  list_name_column = "Phenotype",
  gene_column = "Gene"
)
}
\arguments{
\item{list_name}{The name of the gene list of interest \code{string}}

\item{gene_data}{The data frame of gene list names and associated
genes (data.frame)}

\item{list_name_column}{The name of the column in gene_data that contains
gene list names \code{string}}

\item{gene_column}{The name of the column in gene_data that contains
the genes \code{string}}
}
\value{
A charcter vector of genes associated with the selected list_name
}
\description{
This returns a character vector of genes (a gene list). The gene_data should
be a data frame that contains a column of gene list names (e.g. the column
may be called "Phenotype"), and a column of genes (e.g. "Gene"). For example,
lets call the following data frame \code{phenotype_to_genes}:
}
\details{
\tabular{ll}{
   Phenotype \tab Gene \cr
   "Abnormal heart" \tab gene X \cr
   "Abnormal heart" \tab gene Y \cr
   "Poor vision" \tab gene Z \cr
   "Poor vision" \tab gene Y \cr
   etc... \tab  \cr
}


In the example above, if we wanted to extract genes related to "Abnormal heart",
we would run the following example:
}
\examples{

phenotype_to_genes <- HPOExplorer::load_phenotype_to_genes("phenotype_to_genes.txt")

height_genes <- get_gene_list(list_name = "Abnormality of body height",
                             gene_data = phenotype_to_genes,
                             list_name_column = "Phenotype",
                             gene_column = "Gene")

}
