% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unfinished_list_names.R
\name{get_unfinished_list_names}
\alias{get_unfinished_list_names}
\title{Get unfinished list names}
\usage{
get_unfinished_list_names(list_names, save_dir_tmp)
}
\arguments{
\item{list_names}{character vector of gene list names.}

\item{save_dir_tmp}{Folder to save intermediate results files to
(one file per gene list). Set to \code{NULL} to skip saving temporary files.}
}
\value{
A character vector of list_names that still need to be analysed.
}
\description{
Gets the unfinished gene list names. It reads the file names in the results
directory and uses this to deduce which gene lists
have already been analysed.
This means you can pause the analysis of multiple gene lists and it will not
re-analyse the already completed ones when you start again.
}
\examples{
gene_data <- HPOExplorer::load_phenotype_to_genes()
list_names <- unique(gene_data$Phenotype)[seq_len(3)]
save_dir_tmp <- file.path(tempdir(),"results")
res_files <- ewce_para(ctd = ctd,
                       gene_data = gene_data,
                       list_names = list_names,
                       reps = 10,
                       save_dir_tmp = save_dir_tmp)
unfinished <- get_unfinished_list_names(list_names = list_names,
                                        save_dir_tmp = save_dir_tmp)
}
