% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_gene_lists.R
\name{get_valid_gene_lists}
\alias{get_valid_gene_lists}
\title{get valid gene lists}
\usage{
get_valid_gene_lists(
  ctd,
  list_names,
  list_name_column,
  gene_data,
  annotLevel = 1,
  verbose = TRUE
)
}
\arguments{
\item{ctd}{Cell Type Data List generated using
\link[EWCE]{generate_celltype_data}.}

\item{list_names}{character vector of gene list names.}

\item{list_name_column}{The name of the gene_data column
that has the gene list names.}

\item{gene_data}{data frame of gene list names and genes
(see \link[HPOExplorer]{get_gene_lists}).}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{verbose}{Print messages.}
}
\value{
A character vector of list_names that are associated with a valid
number of genes
}
\description{
Gene lists for EWCE must have 4 or more genes.
These genes must also be present
in the CTD file (see EWCE docs), and must be unique (occasionally a gene list
may contain the same gene repeated. This would cause an error).
}
\details{
This function gets gene lists that have at least four unique genes
that are also present in the CTD file.
}
\keyword{internal}
