% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork_plot_full.R
\name{ggnetwork_plot_full}
\alias{ggnetwork_plot_full}
\title{Create interactive network plot start to finish}
\usage{
ggnetwork_plot_full(
  filters,
  keep_descendants = NULL,
  results = load_example_results(),
  hpo = HPOExplorer::get_hpo(),
  q_threshold = 5e-04,
  effect_threshold = 1,
  columns = HPOExplorer::list_columns(),
  colour_var = "effect",
  size_var = "ontLvl_relative",
  add_ont_lvl_absolute = TRUE,
  add_ont_lvl_relative = TRUE,
  method = c("ggnetwork", "visnetwork"),
  interactive = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{effect_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[KGExplorer]{add_hoverboxes}.}

\item{colour_var}{The column from phenos that you wish
to map to node colour.}

\item{size_var}{Column to scale node size by.}

\item{add_ont_lvl_absolute}{Add the absolute ontology level of each HPO term.
See \link[KGExplorer]{get_ontology_levels} for more details.}

\item{add_ont_lvl_relative}{Add the relative ontology level of each HPO term.
See \link[KGExplorer]{get_ontology_levels} for more details.}

\item{method}{Method to construct plot with.}

\item{interactive}{Make the plot interactive.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[HPOExplorer:make_]{HPOExplorer::make_network_plot}}
  \describe{
    \item{\code{phenos}}{A data.table containing HPO IDs and other metadata.}
    \item{\code{label_var}}{Column containing the label for each node in a graph
(e.g. "hpo_name").}
    \item{\code{show_plot}}{Show the plot.}
    \item{\code{hoverbox_column}}{Name of the new hoverbox column to add.}
    \item{\code{preferred_palettes}}{Preferred palettes to use for each column.}
  }}
}
\value{
A named list of outputs,
including a interactive network plot of the selected subset
of results from RD EWCE analysis.
}
\description{
This puts all the functions together from getting the subset of results
to creating the final interactive plot.
}
\examples{
res <- ggnetwork_plot_full(filters = list(CellType = "Microglia"))
}
