% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork_plot_full.R
\name{ggnetwork_plot_full}
\alias{ggnetwork_plot_full}
\title{Create interactive network plot start to finish}
\usage{
ggnetwork_plot_full(
  cell_type,
  ancestor = NULL,
  results = load_example_results(),
  hpo = HPOExplorer::get_hpo(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  columns = HPOExplorer::list_columns(),
  colour_var = "fold_change",
  size_var = "ontLvl",
  add_ontLvl = TRUE,
  interactive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{cell_type}{The cell type of interest to be plotted.
Can be a single string (e.g. \code{"Astrocytes"}) or a character vector
of multiple cell types (e..g. \code{c("Astrocytes","Microglia")}).
Set to \code{NULL} if you wish to include all cell-types that are available
(after \code{q_threshold} and \code{fold_threshold} filtering).
If >1 cell-type remains, results will be aggregated automatically
such that there is only 1 node per phenotype.}

\item{ancestor}{The ancestor to get all descendants of. If \code{NULL},
returns the entirely ontology.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{columns}{A named vector of columns in \code{phenos}
to add to the hoverdata via \link[HPOExplorer]{make_hoverboxes}.}

\item{colour_var}{Column  to be mapped to node colour.}

\item{size_var}{Column name to be mapped  to node size.}

\item{add_ontLvl}{Add the "ontLvl" (ontology level)
column if not already present.}

\item{interactive}{Make the plot interactive with \link[plotly]{ggplotly}.}

\item{verbose}{Print messages.}
}
\value{
A named list of outputs,
including a interactive network plot of the selected subset
of results from RD EWCE analysis.
}
\description{
This puts all the functions together from getting the subset of results
to creating the final interactive plot.
}
\examples{
res <- ggnetwork_plot_full(cell_type = "Microglia")
}
