% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_not_analysed.R
\name{is_not_analysed}
\alias{is_not_analysed}
\title{Is not analysed?}
\usage{
is_not_analysed(list_names, save_dir_tmp)
}
\arguments{
\item{list_names}{character vector of gene list names.}

\item{save_dir_tmp}{Folder to save intermediate results files to
(one file per gene list). Set to \code{NULL} to skip saving temporary files.}
}
\value{
Vector of TRUE or FALSE.
}
\description{
Takes a gene list name and checks the output results directory to see if that
gene list has been analysed yet.
The \link[MSTExplorer]{ewce_para} function outputs the results
for each gene list in to the results directory (\code{save_dir_tmp})
with name format \emph{list_name.rds}.
}
\keyword{internal}
