% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_CTD.R
\name{load_example_CTD}
\alias{load_example_CTD}
\title{Load example CTD (Descartes)}
\usage{
load_example_CTD()
}
\description{
This loads a example CTD for vignettes and testing purposes. It is Descartes
human cell atlas data at annotation level 1.
}
\examples{
CTD <- load_example_CTD()

}
