% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example_results.R
\name{load_example_results}
\alias{load_example_results}
\title{Load example results}
\source{
\code{
d <-  "~/Desktop/ewce/rare_disease_celltyping_apps/cell_select"
#### Descartes_All_Results_extras ####
f0 <- file.path(d,"data/Descartes_All_Results_extras.rds")
r0 <- readRDS(f0)
data.table::setnames(r0,"list","hpo_name")
r0$HPO_id=NULL
r0 <- HPOExplorer:::fix_hpo_ids(dt=r0)
f0new <- file.path(tempdir(),basename(f0))
saveRDS(r0,file = f0new)
piggyback::pb_upload(file = f0new,
                     tag = "v0.0.1", repo = "neurogenomics/MultiEWCE")
#### Descartes_All_Results ####
f1 <- file.path(d,"data/Descartes_All_Results.rds")
r1 <- readRDS(f1)
data.table::setnames(r1,"list","hpo_name")
r1 <- HPOExplorer:::fix_hpo_ids(dt=r1)
f1new <- file.path(tempdir(),basename(f1))
saveRDS(r1,file = f1new)
piggyback::pb_upload(file = f1new,
                     tag = "v0.0.1", repo = "neurogenomics/MultiEWCE")
#### tabulamuris_merged ####
f2 <- file.path(d,"data/tabulamuris_merged.rds")
r2 <- readRDS(f2)
data.table::setnames(r2,"list","hpo_name")
r2 <- HPOExplorer:::fix_hpo_ids(dt=r2)
f2new <- file.path(tempdir(),basename(f2))
saveRDS(r2,file = f2new)
piggyback::pb_upload(file = f2new,
                     tag = "v0.0.1", repo = "neurogenomics/MultiEWCE")
}

\href{https://github.com/neurogenomics/rare_disease_celltyping/}{
Results located in 'results' folder.}
}
\usage{
load_example_results(
  file = c("Descartes_All_Results_extras.symptoms.rds",
    "Descartes_All_Results_extras.symptoms.full_join.rds",
    "Descartes_All_Results_extras.rds", "gen_overlap.symptoms.filt.rds",
    "tabulamuris_merged.rds"),
  tag = "latest",
  save_dir = tools::R_user_dir(package = "MultiEWCE", which = "cache"),
  force_new = FALSE
)
}
\arguments{
\item{file}{File to load:
\itemize{
\item{"Descartes_All_Results_extras.symptoms.full_join.rds":}{
Contains cell type-phenotype and cell type-symptom (phenotype + disease)
enrichment results merged into one table.
Used the \href{https://descartes.brotmanbaty.org/}{Descartes}
CTD ( annotation level 1).
}
\item{"Descartes_All_Results_extras.rds":}{
Contains cell type-phenotype enrichment results.
Used the \href{https://descartes.brotmanbaty.org/}{Descartes}
CTD ( annotation level 1).
}
\item{"tabulamuris_merged.rds":}{
Contains cell type-phenotype enrichment results.
Used the \href{https://tabula-muris.ds.czbiohub.org/}{Tabula Muris}
CTD.
}
}}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{save_dir}{Where to store the file locally.}

\item{force_new}{Download the file even when a local copy already exists.}
}
\value{
dataframe of enrichment results.
}
\description{
This loads a example of enrichment results from \link[MultiEWCE]{gen_results}
using the Human Phenotype Ontology and a given CellTypeDataset (CTD).
Results were then merged together with  \link[MultiEWCE]{merge_results}.
}
\examples{
res <- load_example_results()
}
