% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_results.R
\name{merge_results}
\alias{merge_results}
\title{merge results}
\usage{
merge_results(results_dir = "results", list_name_column = "phenotype")
}
\arguments{
\item{results_dir}{The filepath to results .rds files}

\item{list_name_column}{The column name for gene list_names (e.g. phenotypes)}
}
\value{
dataframe
}
\description{
merges individual .rds files from the results output directory into one
dataframe.
}
