% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_dendro.R
\name{plot_bar_dendro}
\alias{plot_bar_dendro}
\title{Plot bar dendrogram}
\usage{
plot_bar_dendro(
  results = load_example_results(),
  celltype_col = "cl_name",
  target_branches = get_target_branches(),
  keep_ancestors = names(target_branches),
  hpo = HPOExplorer::get_hpo(),
  cl = get_cl(),
  facets = "ancestor_name",
  add_test_target_celltypes = TRUE,
  add_prop_test = FALSE,
  preferred_palettes = "tol",
  legend.position = "none",
  heights = c(0.3, 1, 0.15, 0.3),
  expand_dendro_x = rep(0.01, 2),
  q_threshold = 0.05,
  show_plot = TRUE,
  save_path = NULL,
  height = 16,
  width = 13
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{celltype_col}{Name of the cell type column in the \code{results}.}

\item{target_branches}{A named list of HPO
branches each matched with CL cell type branches that
correspond to on-target cell types across the two ontologies.}

\item{keep_ancestors}{Only HPO terms that have these ancestors will be kept.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{cl}{Cell Ontology (CL) object from
\code{KGExplorer::get_ontology("cl")}.}

\item{facets}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}

\item{add_test_target_celltypes}{Using the significant phenotype-cell type
association \code{results}, run proportional enrichment tests to
determine whether each cell type is overrepresented in a given HPO branch
relative to all other HPO branches. Overrepresented cell types will be
denoted by "*" above its bar.}

\item{add_prop_test}{Add proportional enrichment results
using \code{run_prop_tests}.}

\item{preferred_palettes}{Preferred palettes to use for each column.}

\item{legend.position}{the default position of legends ("none", "left",
"right", "bottom", "top", "inside")}

\item{heights}{Passed to \link[patchwork]{wrap_plots}.}

\item{expand_dendro_x}{Passed to \link[ggplot2]{scale_x_discrete}
in the cell type dendrogram.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{show_plot}{Print the plot to the console.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{height}{Height of the saved plot.}

\item{width}{Width of the saved plot.}
}
\value{
A bar chart with dendrogram of EWCE results in each cell type.
}
\description{
Create a plot summarising \pkg{MSTExplorer} results as a bar chart
with multiple facets showing selected branches from the
Human Phenotype Ontology (HPO). Also shows a dendrogram of celltype-celltype
relationships using the Cell Ontology (CL).
}
\examples{
results <- load_example_results()
out <- plot_bar_dendro(results = results)
}
