% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_congenital_annotations.R
\name{plot_congenital_annotations}
\alias{plot_congenital_annotations}
\title{Plot congenital annotations}
\usage{
plot_congenital_annotations(
  results,
  gpt_annot = HPOExplorer::gpt_annot_codify(),
  hpo = HPOExplorer::get_hpo(),
  fetal_keywords = c("fetal", "fetus", "primordial", "hESC", "embryonic"),
  celltype_col = "author_celltype",
  x_var = c("fetal_celltype", "fetal_only"),
  remove_annotations = c("varies"),
  keep_descendants = NULL,
  by_branch = FALSE,
  q_threshold = 0.05,
  package = "palettetown",
  palette = "mewtwo",
  proportion.test = TRUE,
  add_baseline = TRUE,
  save_path = NULL,
  ...
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{gpt_annot}{A data.table of GPT annotations.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{fetal_keywords}{A character vector of keywords to identify fetal cell types.}

\item{celltype_col}{The column name of the cell type.}

\item{x_var}{X-axis variable to plot.}

\item{remove_annotations}{A character vector of annotations to remove.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{by_branch}{Use HPO ancestors as the x-axis instead of the
frequency of congenital onset.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{package, palette}{Name of the package from which the given palette is to
be extracted. The available palettes and packages can be checked by running
\code{View(paletteer::palettes_d_names)}.}

\item{proportion.test}{Decides whether proportion test for \code{x} variable is to
be carried out for each level of \code{y}. Defaults to \code{results.subtitle}. In
\code{\link[ggstatsplot:ggbarstats]{ggbarstats()}}, only \emph{p}-values from this test will be displayed.}

\item{add_baseline}{Add a horizontal line showing the proportions
expected by random.}

\item{save_path}{The path to save the plot.}

\item{...}{
  Arguments passed on to \code{\link[ggstatsplot:ggbarstats]{ggstatsplot::ggbarstats}}
  \describe{
    \item{\code{data}}{A data frame (or a tibble) from which variables specified are to
be taken. Other data types (e.g., matrix,table, array, etc.) will \strong{not}
be accepted. Additionally, grouped data frames from \code{{dplyr}} should be
ungrouped before they are entered as \code{data}.}
    \item{\code{x}}{The variable to use as the \strong{rows} in the contingency table. Please
note that if there are empty factor levels in your variable, they will be
dropped.}
    \item{\code{y}}{The variable to use as the \strong{columns} in the contingency table.
Please note that if there are empty factor levels in your variable, they
will be dropped. Default is \code{NULL}. If \code{NULL}, one-sample proportion test
(a goodness of fit test) will be run for the \code{x} variable. Otherwise an
appropriate association test will be run. This argument can not be \code{NULL}
for \code{\link[ggstatsplot:ggbarstats]{ggbarstats()}}.}
    \item{\code{counts}}{The variable in data containing counts, or \code{NULL} if each row
represents a single observation.}
    \item{\code{type}}{A character specifying the type of statistical approach:
\itemize{
\item \code{"parametric"}
\item \code{"nonparametric"}
\item \code{"robust"}
\item \code{"bayes"}
}

You can specify just the initial letter.}
    \item{\code{paired}}{Logical indicating whether data came from a within-subjects or
repeated measures design study (Default: \code{FALSE}).}
    \item{\code{results.subtitle}}{Decides whether the results of statistical tests are
to be displayed as a subtitle (Default: \code{TRUE}). If set to \code{FALSE}, only
the plot will be returned.}
    \item{\code{label}}{Character decides what information needs to be displayed
on the label in each pie slice. Possible options are \code{"percentage"}
(default), \code{"counts"}, \code{"both"}.}
    \item{\code{label.args}}{Additional aesthetic arguments that will be passed to
\code{\link[ggplot2:geom_text]{ggplot2::geom_label()}}.}
    \item{\code{sample.size.label.args}}{Additional aesthetic arguments that will be
passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}.}
    \item{\code{digits}}{Number of digits for rounding or significant figures. May also
be \code{"signif"} to return significant figures or \code{"scientific"}
to return scientific notation. Control the number of digits by adding the
value as suffix, e.g. \code{digits = "scientific4"} to have scientific
notation with 4 decimal places, or \code{digits = "signif5"} for 5
significant figures (see also \code{\link[=signif]{signif()}}).}
    \item{\code{digits.perc}}{Numeric that decides number of decimal places for
percentage labels (Default: \code{0L}).}
    \item{\code{bf.message}}{Logical that decides whether to display Bayes Factor in
favor of the \emph{null} hypothesis. This argument is relevant only \strong{for
parametric test} (Default: \code{TRUE}).}
    \item{\code{ratio}}{A vector of proportions: the expected proportions for the
proportion test (should sum to \code{1}). Default is \code{NULL}, which means the null
is equal theoretical proportions across the levels of the nominal variable.
E.g., \code{ratio = c(0.5, 0.5)} for two levels,
\code{ratio = c(0.25, 0.25, 0.25, 0.25)} for four levels, etc.}
    \item{\code{conf.level}}{Scalar between \code{0} and \code{1} (default: \verb{95\%}
confidence/credible intervals, \code{0.95}). If \code{NULL}, no confidence intervals
will be computed.}
    \item{\code{sampling.plan}}{Character describing the sampling plan. Possible options:
\itemize{
\item \code{"indepMulti"} (independent multinomial; default)
\item \code{"poisson"}
\item \code{"jointMulti"} (joint multinomial)
\item \code{"hypergeom"} (hypergeometric).
For more, see \code{\link[BayesFactor:contingencyTableBF]{BayesFactor::contingencyTableBF()}}.
}}
    \item{\code{fixed.margin}}{For the independent multinomial sampling plan, which
margin is fixed (\code{"rows"} or \code{"cols"}). Defaults to \code{"rows"}.}
    \item{\code{prior.concentration}}{Specifies the prior concentration parameter, set
to \code{1} by default. It indexes the expected deviation from the null
hypothesis under the alternative, and corresponds to Gunel and Dickey's
(1974) \code{"a"} parameter.}
    \item{\code{title}}{The text for the plot title.}
    \item{\code{subtitle}}{The text for the plot subtitle. Will work only if
\code{results.subtitle = FALSE}.}
    \item{\code{caption}}{The text for the plot caption. This argument is relevant only
if \code{bf.message = FALSE}.}
    \item{\code{legend.title}}{Title text for the legend.}
    \item{\code{xlab}}{Label for \code{x} axis variable. If \code{NULL} (default),
variable name for \code{x} will be used.}
    \item{\code{ylab}}{Labels for \code{y} axis variable. If \code{NULL} (default),
variable name for \code{y} will be used.}
    \item{\code{ggtheme}}{A \code{{ggplot2}} theme. Default value is
\code{\link[ggstatsplot:theme_ggstatsplot]{theme_ggstatsplot()}}. Any of the \code{{ggplot2}} themes (e.g.,
\code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}), or themes from extension packages are allowed
(e.g., \code{ggthemes::theme_fivethirtyeight()}, \code{hrbrthemes::theme_ipsum_ps()},
etc.). But note that sometimes these themes will remove some of the details
that \code{{ggstatsplot}} plots typically contains. For example, if relevant,
\code{\link[ggstatsplot:ggbetweenstats]{ggbetweenstats()}} shows details about multiple comparison test as a
label on the secondary Y-axis. Some themes (e.g.
\code{ggthemes::theme_fivethirtyeight()}) will remove the secondary Y-axis and
thus the details as well.}
    \item{\code{package,palette}}{Name of the package from which the given palette is to
be extracted. The available palettes and packages can be checked by running
\code{View(paletteer::palettes_d_names)}.}
    \item{\code{ggplot.component}}{A \code{ggplot} component to be added to the plot prepared
by \code{{ggstatsplot}}. This argument is primarily helpful for \code{grouped_}
variants of all primary functions. Default is \code{NULL}. The argument should
be entered as a \code{{ggplot2}} function or a list of \code{{ggplot2}} functions.}
  }}
}
\description{
Test whether there is a difference in proportion if significantly associated
fetal and non-fetal cell types across phenotypes with congenital onset
vs. those without.
}
\examples{
results <- load_example_results()
results2 <- plot_congenital_annotations(results=results)
}
