% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differential_outcomes_heatmap.R
\name{plot_differential_outcomes_heatmap}
\alias{plot_differential_outcomes_heatmap}
\title{Plot differential outcomes: heatmap}
\usage{
plot_differential_outcomes_heatmap(
  phenotypes = NULL,
  results = load_example_results(),
  celltype_col = "cl_name",
  outcome_var = "AgeOfDeath_score_mean",
  x_var = "celltype_symptom",
  title = "Differential outcomes by cell type",
  subtitle = paste0("Phenotype(s): ", paste(phenotypes, collapse = "; ")),
  fill_limits = NULL,
  print_phenotypes = TRUE,
  show_plot = TRUE,
  save_path = NULL,
  height = NULL,
  width = NULL,
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=add_symptom_results]{add_symptom_results}}
  \describe{
    \item{\code{top_n}}{Top N genes to keep when grouping by \code{group_vars}.}
    \item{\code{results}}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}
    \item{\code{q_threshold}}{The q value threshold to subset the \code{results} by.}
    \item{\code{fold_threshold}}{The minimum fold change in specific expression
to subset the \code{results} by.}
    \item{\code{phenotype_to_genes}}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}
  }}
}
\description{
Plot differential outcomes: heatmap
}
\examples{
results <- load_example_results()
keep_descendants <- "Hypotonia" # HP:0001252
results2 <- HPOExplorer::filter_descendants(results,
                                            keep_descendants = keep_descendants)
results2 <- HPOExplorer::add_death(results2,
                                  allow.cartesian = TRUE,
                                  agg_by = c("disease_id","hpo_id"))
out <- plot_differential_outcomes_heatmap(results=results2)
}
