% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ontology_levels.R
\name{plot_ontology_levels}
\alias{plot_ontology_levels}
\title{Plot ontology levels}
\usage{
plot_ontology_levels(
  results = load_example_results(),
  p2g = HPOExplorer::load_phenotype_to_genes(),
  ctd_list = load_example_ctd(file = paste0("ctd_", unique(results$ctd), ".rds"),
    multi_dataset = TRUE),
  x_vars = c("genes", "cell types", "estimate", "mean_specificity"),
  log_vars = x_vars \%in\% c("estimate", "statistic", "F", "ges", "effect"),
  sig_vars = x_vars \%in\% c("estimate", "statistic", "F", "ges", "effect", "cell types",
    "mean_specificity"),
  group_vars = c("hpo_id", "ontLvl", "ctd"),
  q_threshold = 0.05,
  min_value = NULL,
  label.x.npc = 0.05,
  label.y.npc = 0.5,
  n.breaks = 4,
  notch = FALSE,
  nrow = 2,
  show_plot = TRUE,
  save_path = NULL,
  height = 7,
  width = length(x_vars) * 5.75,
  smooth.line.args = list(method = "loess", se = FALSE),
  return_data = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{p2g}{Phenotype to gene data.}

\item{ctd_list}{A named list of CellTypeDataset objects each
created with \link[EWCE]{generate_celltype_data}.}

\item{x_vars}{Variables to plot on the x-axis of each subplot.}

\item{log_vars}{Logical vector indicating which variables to log-transform.}

\item{sig_vars}{Logical vector indicating which variables to only plot
for significant results.}

\item{group_vars}{Compute the mean value for each \code{x_vars}
grouped by the \code{group_vars}.
This can be less accurate for some metrics but helps to drastically
reduce computational load.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{min_value}{Minimum value for the \code{specificity} metric.}

\item{label.x.npc, label.y.npc}{can be \code{numeric} or \code{character}
 vector of the same length as the number of groups and/or panels. If too
 short they will be recycled. \itemize{ \item If \code{numeric}, value should
 be between 0 and 1. Coordinates to be used for positioning the label,
 expressed in "normalized parent coordinates". \item If \code{character},
 allowed values include: i) one of c('right', 'left', 'center', 'centre',
 'middle') for x-axis; ii) and one of c( 'bottom', 'top', 'center', 'centre',
 'middle') for y-axis.}

 If too short they will be recycled.}

\item{n.breaks}{Passed to \link[ggplot2]{scale_fill_viridis_c}.}

\item{notch}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{nrow}{Number of facet rows for the plot.}

\item{show_plot}{Print the plot to the console.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{height}{Height of the saved plot.}

\item{width}{Width of the saved plot.}

\item{smooth.line.args}{A list of additional aesthetic arguments to be passed
to \code{geom_smooth} geom used to display the regression line.}

\item{return_data}{Return the full long data used in the plots.}
}
\value{
A named list containing the data and the plot.
}
\description{
Generate plots comparing the ontology level of each HPO phenotype and
several other metrics.
}
\examples{
out <- plot_ontology_levels()
}
