% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report.R
\name{plot_report}
\alias{plot_report}
\title{Report plot}
\usage{
plot_report(
  rep_dt,
  results,
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(1),
  annot = HPOExplorer::load_phenotype_to_genes(3),
  remove_cols = c("Rows", "Rows_diff", "ids"),
  label.size = 0.25,
  show_plot = TRUE,
  add_tiers = TRUE,
  save_path = tempfile(fileext = "_plot_report.pdf"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{rep_dt}{Report table.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{phenotype_to_genes}{Phenotype to gene mapping from
\link[HPOExplorer]{load_phenotype_to_genes}.}

\item{annot}{HPO annotations.}

\item{remove_cols}{Columns to remove from \code{rep_dt}.}

\item{label.size}{Size of label border, in mm.}

\item{show_plot}{Print the plot to the console.}

\item{add_tiers}{Add severity tiers subplot.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:ggsave]{ggplot2::ggsave}}
  \describe{
    \item{\code{filename}}{File name to create on disk.}
    \item{\code{plot}}{Plot to save, defaults to last plot displayed.}
    \item{\code{device}}{Device to use. Can either be a device function
(e.g. \link{png}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only). If
\code{NULL} (default), the device is guessed based on the \code{filename} extension.}
    \item{\code{path}}{Path of the directory to save plot to: \code{path} and \code{filename}
are combined to create the fully qualified file name. Defaults to the
working directory.}
    \item{\code{scale}}{Multiplicative scaling factor.}
    \item{\code{width,height}}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}
    \item{\code{units}}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}
    \item{\code{dpi}}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
    \item{\code{limitsize}}{When \code{TRUE} (the default), \code{ggsave()} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
    \item{\code{bg}}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
    \item{\code{create.dir}}{Whether to create new directories if a non-existing
directory is specified in the \code{filename} or \code{path} (\code{TRUE}) or return an
error (\code{FALSE}, default). If \code{FALSE} and run in an interactive session,
a prompt will appear asking to create a new directory when necessary.}
  }}
}
\value{
ggplot object
}
\description{
Plot the results of a filtering report generated by
\link[MSTExplorer]{prioritise_targets}.
}
\examples{
results <- load_example_results()
rep_dt <- example_targets$report
gp <- plot_report(rep_dt=rep_dt, results=results)
}
