% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_celltypes.R
\name{predict_celltypes}
\alias{predict_celltypes}
\title{Predict cell types}
\usage{
predict_celltypes(
  phenotypes,
  diseases_include = NULL,
  diseases_exclude = NULL,
  genes_include = NULL,
  genes_exclude = NULL,
  gene_weights = list(include = 2, default = 1, exclude = 0),
  results = MSTExplorer::load_example_results(),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  agg_var = c("cl_name"),
  effect_var = "logFC",
  x_var = agg_var[1],
  y_var = "score_mean",
  fill_var = "score_sum",
  evidence_score_var = "evidence_score_sum",
  max_x_var = 10,
  subtitle_size = 9,
  plot.margin = ggplot2::margin(1, 1, 1, 40),
  show_plot = TRUE,
  save_path = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{phenotypes}{Phenotypes observed in the patient.
Can be a list of HPO phenotype IDs or HPO phenotype names.}

\item{diseases_include}{Diseases that the patient is known to have.
Can be provided as OMIM, Orphanet, or DECIPHER disease IDs.}

\item{diseases_exclude}{Diseases that the patient is known NOT to have.
Can be provided as OMIM, Orphanet, or DECIPHER disease IDs.}

\item{genes_include}{Genes in which the patient is known to have
abnormalities.}

\item{genes_exclude}{Genes in which the patient is known NOT to have
abnormalities.}

\item{gene_weights}{A named list describing the weight to apply to
genes in the include, default, and exclude lists.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{phenotype_to_genes}{Phenotype to gene mapping from
\link[HPOExplorer]{load_phenotype_to_genes}.}

\item{agg_var}{The variable(s) to aggregate \code{results} by.}

\item{effect_var}{Name of the effect size column in the \code{results}.}

\item{x_var}{Variable to plot on the x-axis.}

\item{y_var}{Variable to plot on the y-axis.}

\item{fill_var}{Variable to fill by.}

\item{evidence_score_var}{Which variable from
\link[HPOExplorer]{add_evidence} to use when weighting genes.}

\item{max_x_var}{The maximum number of cell types to display.}

\item{subtitle_size}{Size of the plot subtitle.}

\item{plot.margin}{margin around entire plot (\code{unit} with the sizes of
the top, right, bottom, and left margins); inherits from \code{margin}.}

\item{show_plot}{Print the plot to the console.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{width}{Width of the saved plot.}

\item{height}{Height of the saved plot.}
}
\value{
\link[data.table]{data.table} of prioritised cell types,
sorted by a "score" that combines:
\itemize{
\item{The phenotype-cell type enrichment p-values ("p").}
\item{The phenotype-cell type enrichment effect size ("effect").}
\item{A gene-wise factor that upweights/downweights
included/excluded genes respectively,
multiplied by the evidence score of a phenotype-gene association.
Only applied when \code{genes_include} or \code{genes_exclude} is provided. }
}
}
\description{
Predict the causal cell types underlying a patient's phenotypes given some
varying degree of prior knowledge.
}
\examples{
phenotypes <- c("Generalized neonatal hypotonia",
                "Scrotal hypospadias",
                "Increased circulating progesterone")
# diseases_include <- "OMIM:176270"
genes_include <- c("MAGEL2","HERC2")
genes_exclude <- c("SNORD115-1")
ct <- predict_celltypes(phenotypes = phenotypes,
                        genes_include = genes_include,
                        genes_exclude = genes_exclude)
}
