% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritise_targets.R
\name{prioritise_targets}
\alias{prioritise_targets}
\title{Prioritise target genes}
\usage{
prioritise_targets(
  results = load_example_results(),
  ctd_list = load_example_ctd(c("ctd_DescartesHuman.rds", "ctd_HumanCellLandscape.rds"),
    multi_dataset = TRUE),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  hpo = HPOExplorer::get_hpo(),
  keep_deaths = HPOExplorer::list_deaths(exclude = c("Miscarriage", "Stillbirth",
    "Prenatal death"), include_na = TRUE),
  keep_descendants = c("Phenotypic abnormality"),
  keep_ont_levels = NULL,
  pheno_ndiseases_threshold = NULL,
  gpt_filters = NULL,
  severity_score_gpt_threshold = 20,
  keep_tiers = NULL,
  severity_threshold_max = NULL,
  info_content_threshold = 8,
  run_prune_ancestors = TRUE,
  severity_threshold = NULL,
  pheno_frequency_threshold = NULL,
  keep_onsets = HPOExplorer::list_onsets(include_na = TRUE),
  effect_var = "logFC",
  q_threshold = 0.05,
  effect_threshold = 1,
  symptom_intersection_threshold = 0.25,
  keep_celltypes = NULL,
  evidence_score_threshold = 15,
  keep_chr = c(seq(22), "X", "Y"),
  gene_size = list(min = 0, max = Inf),
  gene_frequency_threshold = NULL,
  keep_biotypes = NULL,
  keep_specificity_quantiles = seq(30, 40),
  keep_mean_exp_quantiles = seq(30, 40),
  sort_cols = c(severity_score_gpt = -1, q = 1, logFC = -1, specificity = -1, mean_exp =
    -1, pheno_freq_mean = -1, gene_freq_mean = -1, width = 1),
  top_n = NULL,
  group_vars = c("hpo_id"),
  return_report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{ctd_list}{A named list of CellTypeDataset objects each
created with \link[EWCE]{generate_celltype_data}.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{keep_deaths}{The age of death associated with each HPO ID to keep.
If >1 age of death is associated with the term,
only the earliest age is considered.
See \link[HPOExplorer]{add_death} for details.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{keep_ont_levels}{Only keep phenotypes at certain \emph{absolute}
ontology levels to keep.
See \link[HPOExplorer]{add_ont_lvl} for details.}

\item{pheno_ndiseases_threshold}{Filter phenotypes by the
maximum number of diseases they are associated with.}

\item{gpt_filters}{A named list of filters to apply to the GPT annotations.}

\item{severity_score_gpt_threshold}{The minimum GPT severity score that a
phenotype can have across any disease.}

\item{keep_tiers}{Tiers from \link[HPOExplorer]{hpo_tiers} to keep.
Include \code{NA} if you wish to retain phenotypes that
do not have any Tier assignment.}

\item{severity_threshold_max}{The max severity score that a phenotype can
have across any disease.}

\item{info_content_threshold}{Minimum phenotype information content
threshold.}

\item{run_prune_ancestors}{Prune redundant ancestral terms if any of their
descendants are present. Passes to \link[KGExplorer]{prune_ancestors}.}

\item{severity_threshold}{Only keep phenotypes with a mean
severity score (averaged across multiple associated diseases) below the
set threshold. The severity score ranges from 1-4 where 1 is the MOST severe.
Include \code{NA} if you wish to retain phenotypes that
do not have any severity score.}

\item{pheno_frequency_threshold}{Only keep phenotypes with frequency
above the set threshold. Frequency ranges from 0-100 where 100 is
a phenotype that occurs 100\% of the time in all associated diseases.
Include \code{NA} if you wish to retain phenotypes that
do not have any frequency data.
See \link[HPOExplorer]{add_pheno_frequency} for details.}

\item{keep_onsets}{The age of onset associated with each HPO ID to keep.
If >1 age of onset is associated with the term,
only the earliest age is considered.
See \link[HPOExplorer]{add_onset} for details.}

\item{effect_var}{Name of the effect size column in the \code{results}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{effect_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{symptom_intersection_threshold}{Minimum proportion of genes
overlapping between a symptom gene list
(phenotype-associated genes in the context of a particular disease)
and the phenotype-cell type association driver genes}

\item{keep_celltypes}{Cell type to keep.}

\item{evidence_score_threshold}{The minimum threshold of mean
evidence scores of each gene-phenotype association to keep.}

\item{keep_chr}{Chromosomes to keep.}

\item{gene_size}{Min/max gene size (important for therapeutics design).}

\item{gene_frequency_threshold}{Only keep genes with frequency
above the set threshold. Frequency ranges from 0-100 where 100 is
a gene that occurs 100\% of the time in a given phenotype.
Include \code{NA} if you wish to retain genes that
do not have any frequency data.
See \link[HPOExplorer]{add_gene_frequency} for details.}

\item{keep_biotypes}{Which gene biotypes to keep.
(e.g. "protein_coding", "processed_transcript", "snRNA",
"lincRNA", "snoRNA", "IG_C_gene")}

\item{keep_specificity_quantiles}{Which cell type
specificity quantiles to keep (max quantile is 40).}

\item{keep_mean_exp_quantiles}{Which cell type
mean expression quantiles to keep (max quantile is 40).}

\item{sort_cols}{How to sort the rows using \link[data.table]{setorderv}.
\code{names(sort_cols)} will be supplied to the \code{cols=} argument
and values will be supplied to the \code{order=} argument.}

\item{top_n}{Top N genes to keep when grouping by \code{group_vars}.}

\item{group_vars}{Columns to group by when selecting \code{top_n} genes.}

\item{return_report}{If \code{TRUE}, will return a named list containing a
\code{report} that shows the number of
phenotypes/celltypes/genes remaining after each filtering step.}

\item{verbose}{Print messages.}
}
\value{
A data.table of the prioritised phenotype- and
cell type-specific gene targets.
}
\description{
Prioritise target genes based on a procedure:\cr
\enumerate{
\item{Disease-level: \code{keep_deaths}: }{
Keep only diseases with a certain age of death.}
\item{Disease-level: \code{severity_threshold_max}: }{
\preformatted{
Keep only diseases annotated as a certain degree of severity or greater
 (filters on maximum severity per disease).}
}
\item{Phenotype-level: \code{prune_ancestors}: }{
\preformatted{
Remove redundant ancestral phenotypes when at least one of their
 descendants already exist.}
}
\item{Phenotype-level: \code{keep_descendants}: }{
\preformatted{
Remove phenotypes belonging to a certain branch of the HPO,
 as defined by an ancestor term.}
}
\item{Phenotype-level: \code{keep_ont_levels}: }{
Keep only phenotypes at certain absolute ontology levels within the HPO.}
\item{Phenotype-level: \code{pheno_ndiseases_threshold}: }{
The maximum number of diseases each phenotype can be associated with.}
\item{Phenotype-level: \code{keep_tiers}: }{
Keep only phenotypes with high severity Tiers.}
\item{Phenotype-level: \code{severity_threshold}: }{
Keep only phenotypes with mean Severity equal to or below the threshold.}
\item{Phenotype-level: \code{gpt_filters}: }{
\preformatted{
Keep only phenotypes with certain GPT annotations in specific
 severity metrics.}
}
\item{Phenotype-level: \code{severity_score_gpt_threshold}: }{
Keep only phenotypes with a minimum GPT severity score.}
\item{Phenotype-level: \code{info_content_threshold}: }{
\preformatted{
Keep only phenotypes with a minimum information criterion score
 (computed from the HPO).}
}
\item{Symptom-level: \code{pheno_frequency_threshold}: }{
\preformatted{
Keep only phenotypes with mean frequency equal to or above the threshold
 (i.e. how frequently a phenotype is associated with any diseases in
 which it occurs).}
}
\item{Symptom-level: \code{keep_onsets}: }{
Keep only symptoms with a certain age of onset.}
\item{Symptom-level: \code{symptom_p_threshold}: }{
Uncorrected p-value threshold to filter cell type-symptom associations by.}
\item{Symptom-level: \code{symptom_intersection_threshold}: }{
\preformatted{
Minimum proportion of genes overlapping between a symptom gene list
 (phenotype-associated genes in the context of a particular disease)
 and the phenotype-cell type association driver genes.}
}
\item{Cell type-level: \code{q_threshold}: }{
\preformatted{
Keep only cell type-phenotype association results at q<=0.05.}
}
\item{Cell type-level: \code{effect_threshold}: }{
Keep only cell type-phenotype association results at effect size>=1.}
\item{Cell type-level: \code{keep_celltypes}: }{
Keep only terminally differentiated cell types.}
\item{Gene-level: \code{keep_chr}: }{
Remove genes on non-standard chromosomes.}
\item{Gene-level: \code{evidence_score_threshold}: }{
\preformatted{
Remove genes that are below an aggregate phenotype-gene
 evidence score threshold.}
}
\item{Gene-level: \code{gene_size}: }{
Keep only genes <4.3kb in length.}
\item{Gene-level: \code{add_driver_genes}: }{
\preformatted{
Keep only genes that are driving the association with a given phenotype
 (inferred by the intersection of phenotype-associated genes and gene with
 high-specificity quantiles in the target cell type).}
}
\item{Gene-level: \code{keep_biotypes}: }{
Keep only genes belonging to certain biotypes.}
\item{Gene-level: \code{gene_frequency_threshold}: }{
\preformatted{
Keep only genes at or above a certain mean frequency threshold
 (i.e. how frequently a gene is associated with a given phenotype
 when observed within a disease).}
}
\item{Gene-level: \code{keep_specificity_quantiles}: }{
\preformatted{
Keep only genes in top specificity quantiles
 from the cell type dataset (CTD).}
}
\item{Gene-level: \code{keep_mean_exp_quantiles}: }{
\preformatted{
Keep only genes in top mean expression quantiles
 from the cell type dataset (CTD).}
}
\item{Gene-level: \code{symptom_gene_overlap}: }{
\preformatted{
Ensure that genes nominated at the phenotype-level also
 appear in the genes overlapping at the cell type-specific symptom-level.}
}
\item{All levels: \code{sort_cols}: }{
\preformatted{
Sort candidate targets by one or more columns
 (e.g. "severity_score_gpt", "q").
}
}
\item{All levels: \code{top_n}: }{
\preformatted{
Only return the top N targets per variable group
 (specified with the "group_vars" argument).
 For example, setting "group_vars" to "hpo_id" and "top_n" to 1 would
 only return one target (row) per phenotype ID after sorting.}
}
}
}
\details{
Term key:\cr
\itemize{
\item{Disease: }{
\preformatted{
A disease defined in the database
OMIM, DECIPHER and/or Orphanet.}
}
\item{Phenotype: }{A clinical feature associated with one or more diseases.}
\item{Symptom: }{
\preformatted{
A phenotype within the context of a particular disease.
Within a given phenotype, there may be multiple symptoms with
 partially overlapping genetic mechanisms.}
}
\item{Assocation: }{
\preformatted{
A cell type-specific enrichment test result conducted
at the disease-level, phenotype-level, or symptom-level.}
}
}
}
\examples{
results = load_example_results()[q<0.05]
out <- prioritise_targets(results=results)
}
