% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritise_targets.R
\name{prioritise_targets}
\alias{prioritise_targets}
\title{Prioritise target genes}
\usage{
prioritise_targets(
  results,
  ctd,
  annotLevel = 1,
  q_threshold = 0.05,
  fold_threshold = 1,
  keep_celltypes = terminal_celltypes()$CellType,
  keep_onsets = HPOExplorer::list_onsets(postnatal_only = TRUE),
  keep_tiers = 1,
  gene_size = list(min = 0, max = 4300),
  keep_seqnames = c(seq_len(22), "X", "Y"),
  keep_biotypes = NULL,
  keep_specificity_quantiles = seq(38, 40),
  sort_cols = c(tier = 1, q = 1, fold_change = -1, specificity_quantile = -1, specificity
    = -1, mean_exp = -1, width = 1),
  top_n = 3,
  group_vars = c("HPO_ID", "CellType"),
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes(),
  hpo = HPOExplorer::get_hpo(),
  verbose = TRUE
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}.}

\item{ctd}{Cell Type Data List generated using
\link[EWCE]{generate_celltype_data}.}

\item{annotLevel}{An integer indicating which level of \code{sct_data} to
analyse (\emph{Default: 1}).}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{keep_celltypes}{Cell type to keep.}

\item{keep_onsets}{The age of onset associated with each HPO ID to keep.}

\item{keep_tiers}{Tiers from \link[HPOExplorer]{hpo_tiers} to keep.}

\item{keep_seqnames}{Chromosomes to keep.}

\item{sort_cols}{How to sort the rows using \link[data.table]{setorderv}.
\code{names(sort_cols)} will be supplied to the \code{cols=} argument
and values will be supplied to the \code{order=} argument.}

\item{top_n}{Top N genes to keep when grouping by \code{group_vars}.}

\item{group_vars}{Columns to group by when selecting \code{top_n} genes.}

\item{phenotype_to_genes}{Output of
\link[HPOExplorer]{load_phenotype_to_genes} mapping phenotypes
to gene annotations.}

\item{hpo}{Human Phenotype Ontology object, loaded from \pkg{ontologyIndex}.}

\item{verbose}{Print messages.}
}
\description{
Prioritise target genes based on a procedure:
\enumerate{
\item{Association-level: }{Keep only results at q<=0.05.}
\item{Association-level: }{Keep only results at fold_change>=1.}
\item{Phenotype-level: }{Keep only phenotypes with high severity Tiers}
\item{Phenotype-level: }{Keep only phenotypes with postnatal age of onsets.}
\item{Cell type-level: }{Keep only terminally differentiated cell types.}
\item{Gene-level: }{Keep only genes <4.7kb in length.}
\item{Gene-level: }{Sort genes by cell type-specificity.}
}
}
\examples{
results <- load_example_results()
ctd <- load_example_ctd()
top_targets <- prioritise_targets(results = results,
                                  ctd = ctd)
}
