% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritise_targets_network.R
\name{prioritise_targets_network}
\alias{prioritise_targets_network}
\title{Prioritise targets network}
\source{
\href{https://book.archnetworks.net/visualization}{
book.archnetworks.net}

\href{https://r-graph-gallery.com/310-custom-hierarchical-edge-bundling}{
Edge bundling with \pkg{ggraph}
}
}
\usage{
prioritise_targets_network(
  top_targets,
  vertex_vars = c("Phenotype", "CellType", "Gene"),
  group_var = "ancestor_name",
  edge_var = "fold_change",
  show_plot = TRUE,
  save_path = tempfile(fileext = "_prioritise_targets_network.html"),
  layout = "layout_with_kk",
  randomSeed = 2023,
  verbose = TRUE
)
}
\arguments{
\item{top_targets}{output of \link[MultiEWCE]{prioritise_targets}.}

\item{vertex_vars}{Columns within \code{top_targets}
to include as vertices/nodes within the network.}

\item{show_plot}{Print the plot after it has been created.}

\item{save_path}{Path to save HTML version of network to.}

\item{layout}{: Character Name of igraph layout function to use. Default to "layout_nicely"}

\item{randomSeed}{: Number. The nodes are randomly positioned initially. This means that the settled result is different every time. If you provide a random seed manually, the layout will be the same every time.}

\item{verbose}{Print messages.}
}
\value{
A named list containing the \link[visNetwork]{visNetwork} plot
and the the graph used to make the plot.
}
\description{
Plot the output of \link[MultiEWCE]{prioritise_targets}
as an interactive network with \link[visNetwork]{visNetwork}.
See \link[igraph]{layout} for a list of all layout functions
that can be passed to the \code{prioritise_targets_network(layout=)} argument
as a string (e.g. "layout.fruchterman.reingold" or "layout_as_tree").
See \href{https://rpubs.com/JanpuHou/337696}{here} for some visualized
examples of each layout.
}
\examples{
top_targets <- prioritise_targets()
vn <- prioritise_targets_network(top_targets = top_targets)
}
