% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_phenomix.R
\name{run_phenomix}
\alias{run_phenomix}
\title{Run phenomix}
\usage{
run_phenomix(
  ctd_name,
  annotLevel,
  ymat,
  test_method = "glm",
  metric = "specificity",
  save_path = here::here("results", paste0("phenomix_", test_method, "_", metric),
    paste0("phenomix_", ctd_name, "_results.tsv.gz")),
  workers = NULL,
  force_new = FALSE,
  ...
)
}
\description{
Run phenomix
}
\examples{
ymat <- HPOExplorer::hpo_to_matrix(formula = "gene_symbol ~ hpo_id")
lm_res <- run_phenomix(ctd_name = "HumanCellLandscape",
                       annotLevel = 3,
                       ymat = ymat,
                       save_path=tempfile())
}
