% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_phenos.R
\name{subset_phenos}
\alias{subset_phenos}
\title{Subset RD EWCE results}
\usage{
subset_phenos(
  filters = NULL,
  keep_descendants = NULL,
  results = load_example_results(),
  hpo = HPOExplorer::get_hpo(),
  q_threshold = 5e-04,
  fold_threshold = 1,
  verbose = TRUE
)
}
\arguments{
\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{keep_descendants}{Terms whose descendants should be kept 
(including themselves).
 Set to \code{NULL} (default) to skip this filtering step.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MultiEWCE]{gen_results}
and merged together with \link[MultiEWCE]{merge_results}}

\item{hpo}{Human Phenotype Ontology object,
loaded from \link[KGExplorer]{get_ontology}.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{fold_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{verbose}{Print messages.}
}
\value{
A data frame of results taken from the main data frame of results.
}
\description{
This subsets  the Rare disease EWCE results by cell type,
q threshold and fold change.
}
\examples{
phenos <- subset_phenos(filters = list(CellType = "Cardiomyocytes"),
                        keep_descendants = "Neurodevelopmental delay")
}
