% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_results.R
\name{subset_results}
\alias{subset_results}
\title{Subset EWCE results}
\usage{
subset_results(
  filters = list(cl_name = NULL),
  results = load_example_results(),
  q_threshold = 5e-04,
  effect_threshold = 1,
  verbose = TRUE
)
}
\arguments{
\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{effect_threshold}{The minimum fold change in specific expression
to subset the \code{results} by.}

\item{verbose}{Print messages.}
}
\value{
A data frame of the selected subset of RD EWCE results
with HPO ID column added.
}
\description{
Subset RD EWCE results data by cell type, fold change and q value.
}
\examples{
signif_cell_data <- subset_results(filters=list(CellType = "Cardiomyocytes"))
}
