% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_results.R
\name{summarise_results}
\alias{summarise_results}
\title{Summarise results}
\usage{
summarise_results(
  results,
  group_var = "ctd",
  digits = 3,
  save_path = tempfile("summarise_results.csv")
)
}
\description{
Summarise results from \pkg{MSTExplorer}.
}
