% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_target_celltypes.R
\name{test_target_celltypes}
\alias{test_target_celltypes}
\title{Test target cell types}
\usage{
test_target_celltypes(
  results = load_example_results(),
  target_celltypes = get_target_celltypes(),
  celltype_col = "cl_id",
  tests = c("within_branches", "across_branches", "across_branches_per_celltype"),
  method = c("glm", "anova"),
  within_var = "hpo_id",
  ancestor_var = "ancestor_name",
  q_threshold = 0.05,
  cores = 1
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{target_celltypes}{A named list of HPO branches and matched CL ids.}

\item{celltype_col}{Name of the cell type column in the \code{results}.}

\item{tests}{The types of tests to conduct.
\describe{
\item{"within_branches"}{
Are on-target cell types enriched for significant results?}
\item{"across_branches"}{
Are on-target cell types enriched for significant results across branches?}
\item{"across_branches_per_celltype"}{
Are on-target cell types enriched for significant results across
branches per cell type?}
}}

\item{method}{Methods to run stats with.}

\item{within_var}{Within-subject variable.}

\item{ancestor_var}{The name of the \code{results} column
containing the ancestor name.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}

\item{cores}{Number of CPU cores to parallelise the
proportional enrichment tests across.}
}
\description{
Test the following hypotheses across \link{gen_results} enrichment results.
}
\examples{
res <- test_target_celltypes(tests="within_branches")
}
