% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttd_check.R
\name{ttd_check}
\alias{ttd_check}
\title{Gene therapy targets}
\usage{
ttd_check(
  top_targets,
  drug_types = NULL,
  failed_status = c("Terminated", "Withdrawn from market", "Discontinued*", NA),
  keep_status = NULL,
  remove_status = c(NA),
  allow.cartesian = FALSE,
  run_map_genes = TRUE,
  force_new = FALSE,
  show_plot = TRUE,
  save_path = NULL,
  height = NULL,
  width = NULL,
  phenotype_to_genes = HPOExplorer::load_phenotype_to_genes()
)
}
\arguments{
\item{top_targets}{Top targets generated by
\link[MSTExplorer]{prioritise_targets}.}

\item{drug_types}{Filter results by drug type.}

\item{failed_status}{Drug approval status categories that indicate
the drug failed.}

\item{keep_status}{Filter results by drug approval status.}

\item{remove_status}{Remove results by drug approval status.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{run_map_genes}{Map genes to standardised HGNC symbols using 
\link[orthogene]{map_genes}.}

\item{force_new}{If TRUE, force a new download.}

\item{show_plot}{Print the plot to the console.}

\item{save_path}{Save the plot to a file.
Set to \code{NULL} to not save the plot.}

\item{height}{Height of the saved plot.}

\item{width}{Width of the saved plot.}

\item{phenotype_to_genes}{Phenotype to gene mapping from
\link[HPOExplorer]{load_phenotype_to_genes}.}
}
\description{
Identify the overlap between your prioritised list of gene therapy targets
and currently existing gene therapy targets that are currently on the market
or are in clinical trials. Uses data from the
\href{https://db.idrblab.net/ttd/full-data-download}{
Therapuetic Target Database}.
}
\examples{
top_targets <- MSTExplorer::example_targets$top_targets
res <- ttd_check(top_targets=top_targets)
}
