% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_associations_correlate_ctd.R
\name{validate_associations_correlate_ctd}
\alias{validate_associations_correlate_ctd}
\title{Validate associations: correlate CellTypeDataset}
\usage{
validate_associations_correlate_ctd(
  results = load_example_results(),
  filters = NULL,
  group_var = "ctd",
  celltype_var = "cl_name",
  q_threshold = 0.05
)
}
\arguments{
\item{results}{The cell type-phenotype enrichment results generated by
\link[MSTExplorer]{gen_results}
and merged together with \link[MSTExplorer]{merge_results}}

\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{group_var}{A character string specifying the column in \code{results}
that contains the group variable to compare across.}

\item{celltype_var}{A character string specifying the column in \code{results}
that contains the cell type variable to compare across.}

\item{q_threshold}{The q value threshold to subset the \code{results} by.}
}
\description{
Validate phenotype-cell type associations can making comparisons within and
between results from different CellTypeDatasets.
}
\examples{
results <- load_example_results()[,.SD[seq(10000)],by="ctd"]
#### Across CTD ####
out1 <- validate_associations_correlate_ctd(results=results,
                                            group_var="ctd")
#### Within CTD: across developmental stages ####
filters <- list(ctd=c("HumanCellLandscape"), stage=c("Fetus","Adult"))
out2 <- validate_associations_correlate_ctd(results=results,
                                            filters=filters,
                                            group_var="stage")
}
